refactor(cli): carve all top-level commands into cli/commands/
Final decomposition step: move the 18 top-level commands out of the
monolith into one module each under cli/commands/ (demo, patterns, packs,
validate, check, explain, replay, report, serve, scan, export,
export_sessions, eval, init, doctor, onboard, mode, prompt). Command-local
helpers travel with their command (scan's yaml filters, export's OTLP
converters, mode's yaml patcher); cross-command coupling is explicit
(onboard imports scan/mode helpers + host's mode resolver; export_sessions
imports export's OTLP converters).

_monolith.py is deleted. __init__ imports every command/group module (for
registration) and re-exports the full public + internal API, so
'from sponsio.cli import X' is unchanged. main() now lives in __init__.

Command set identical (23 top-level + subgroups); full suite 2296 passed;
python -m sponsio.cli and the console script both work.
