--- sponsio/cli/_monolith.py
@@ -1,4400 +1,0 @@
-"""Sponsio CLI entry point."""
-
-from __future__ import annotations
-
-import json
-import os
-import re
-import subprocess
-import sys
-import urllib.error
-import urllib.request
-from pathlib import Path
-
-import click
-
-from sponsio.cli.app import cli
-
-from sponsio.cli._shared import (
-    _contract_guarantee,
-    _looks_like_sponsio_config,
-    _parse_existing_contracts,
-    _parse_since,
-    _resolve_entry,
-)
-
-# onboard reuses the host group's runtime-mode resolver.
-from sponsio.cli.groups.host import _resolve_runtime_mode
-from sponsio.constants import DASHBOARD_DEFAULT_PORT
-
-
-# ---------------------------------------------------------------------------
-# demo
-# ---------------------------------------------------------------------------
-
-
-@cli.command()
-@click.option(
-    "--scenario",
-    default="cleanup",
-    type=click.Choice(["cleanup", "backup", "wire", "freeze"], case_sensitive=False),
-    help="Demo scenario: cleanup (default), backup, wire, freeze",
-)
-@click.option(
-    "--mode",
-    default="mock",
-    type=click.Choice(["mock", "integration"], case_sensitive=False),
-    show_default=True,
-    help="mock uses no optional SDKs; integration runs repo example scripts.",
-)
-@click.option("--no-guard", is_flag=True, help="Replay the unsafe trajectory.")
-@click.option("--fast", is_flag=True, help="Skip typing delays.")
-def demo(scenario: str, mode: str, no_guard: bool, fast: bool):
-    """Run a Sponsio demo in your terminal.
-
-    Four trajectory replays showing unsafe agent behavior and the
-    contracts that block it. The default mock mode works from a plain
-    PyPI install with no API key and no optional framework SDKs.
-
-    \b
-      cleanup . Claude Code cleanup agent deletes `.env` & `.git/`
-      backup  . SRE cost-optimizer deletes prod DR backups (OWASP ASI-10)
-      wire    . AP copilot wires $847k to an unverified vendor (OWASP ASI-09)
-      freeze  . Replit-style agent violates code freeze + hides it (OWASP ASI-10)
-
-    Examples:\n
-        sponsio demo\n
-        sponsio demo --scenario freeze --fast\n
-        sponsio demo --scenario wire --no-guard\n
-        sponsio demo --mode integration --scenario freeze
-    """
-    scenario_map = {
-        "cleanup": ("demo_coding_cleanup.py", "Coding Agent \u2014 Cleanup gone rogue"),
-        "backup": (
-            "demo_backup_delete.py",
-            "SRE Cost-Optimizer \u2014 Prod DR backups deleted",
-        ),
-        "wire": (
-            "demo_wire_transfer.py",
-            "AP Copilot \u2014 Fraudulent wire transfer",
-        ),
-        "freeze": (
-            "demo_freeze_violation.py",
-            "Coding Agent \u2014 Code-freeze violation + coverup",
-        ),
-    }
-
-    script_name, label = scenario_map[scenario]
-
-    click.echo()
-    click.echo(click.style("Sponsio Demo", bold=True))
-    click.echo(click.style(f"  {label}", fg="cyan"))
-    click.echo()
-
-    if mode == "mock":
-        from sponsio.demos.replay import run_demo
-
-        run_demo(scenario, no_guard=no_guard, fast=fast)
-        return
-
-    import sponsio
-
-    # Resolve relative to the installed package, not this file's depth,
-    # so it stays correct regardless of where the CLI code lives.
-    repo_root = Path(sponsio.__file__).resolve().parent.parent
-    script_path = repo_root / "examples" / "demo" / script_name
-
-    if not script_path.exists():
-        click.echo(
-            click.style(
-                "Error: integration demo scripts are only available from a "
-                "source checkout. Use the default mock mode from PyPI: "
-                f"{click.style('sponsio demo', bold=True)}",
-                fg="red",
-            )
-        )
-        sys.exit(1)
-
-    try:
-        cmd = [sys.executable, str(script_path)]
-        if no_guard:
-            cmd.append("--no-guard")
-        if fast:
-            cmd.append("--fast")
-        subprocess.run(cmd, check=True)
-    except subprocess.CalledProcessError as e:
-        sys.exit(e.returncode)
-    except KeyboardInterrupt:
-        click.echo("\nInterrupted.")
-
-
-# ---------------------------------------------------------------------------
-# patterns
-# ---------------------------------------------------------------------------
-
-
-@cli.command()
-def patterns():
-    """List all available contract patterns with examples."""
-
-    def _section(title, items, color):
-        click.echo(click.style(title, bold=True))
-        click.echo()
-        for name, example, meaning in items:
-            click.echo(click.style(f"  {name}", fg=color, bold=True))
-            click.echo(f"    Example : {example}")
-            click.echo(click.style(f"    Meaning : {meaning}", dim=True))
-            click.echo()
-
-    # --- Core temporal (14) ---
-    click.echo()
-    _section(
-        "Core Temporal Patterns (14 det)",
-        [
-            ("must_precede", "tool `A` must precede `B`", "A must happen before B"),
-            (
-                "always_followed_by",
-                "tool `A` must always be followed by `B`",
-                "whenever A, eventually B",
-            ),
-            ("no_reversal", "cannot `B` after `A`", "A commits; B forbidden after"),
-            (
-                "requires_permission",
-                "tool `X` requires permission `perm`",
-                "tool needs authorization",
-            ),
-            ("no_data_leak", "no data leak from `src` to `ext`", "data containment"),
-            (
-                "mutual_exclusion",
-                "`A` and `B` are mutually exclusive",
-                "at most one per session",
-            ),
-            ("rate_limit", "tool `X` at most N times", "frequency cap"),
-            ("idempotent", "tool `X` must execute at most once", "single execution"),
-            (
-                "deadline",
-                "`action` within N steps of `trigger`",
-                "time-bounded obligation",
-            ),
-            ("must_confirm", "tool `X` requires confirmation", "human-in-the-loop"),
-            ("cooldown", "N steps between consecutive `X`", "minimum interval"),
-            (
-                "segregation_of_duty",
-                "review and approve by different agents",
-                "separation of concerns",
-            ),
-            ("bounded_retry", "tool `X` limited to N retries", "retry cap"),
-            (
-                "loop_detection",
-                "tool `X` at most N consecutive calls",
-                "runaway loop prevention",
-            ),
-        ],
-        "cyan",
-    )
-
-    # --- Argument / path / length (5) ---
-    _section(
-        "Argument & Path Constraints (5 det)",
-        [
-            (
-                "arg_blacklist",
-                "tool `bash` arg `command` must not match `rm -rf`",
-                "forbid patterns in args",
-            ),
-            (
-                "arg_allowlist",
-                "tool `send_money` arg `recipient` must be one of `US-internal-001`, `US-internal-002`",
-                "arg must match one of the allowed patterns",
-            ),
-            (
-                "scope_limit",
-                "tool `file_write` restricted to `/app/data`",
-                "restrict tool to allowed paths",
-            ),
-            (
-                "arg_length_limit",
-                "tool `bash` arg `command` max 500 chars",
-                "block code-injection via long args",
-            ),
-            (
-                "data_intact",
-                "`grep` must use only original data files",
-                "tool must use unmodified data",
-            ),
-        ],
-        "cyan",
-    )
-
-    # --- OWASP Agentic Top 10 (8) ---
-    _section(
-        "OWASP Agentic Security Patterns (8 det)",
-        [
-            (
-                "destructive_action_gate",
-                "`delete_db` requires approval from `approver`",
-                "human approval + role for destructive ops",
-            ),
-            (
-                "untrusted_source_gate",
-                "after `web_fetch`, `send_email` requires re-confirmation",
-                "re-confirm after untrusted input (A,E pair)",
-            ),
-            (
-                "required_steps_completion",
-                "every `start_task` must be followed by all of [`log`, `notify`]",
-                "all steps must follow trigger",
-            ),
-            (
-                "tool_allowlist",
-                "only [`read_file`, `write_file`] may be called",
-                "first-line defense against injected tools",
-            ),
-            (
-                "dangerous_bash_commands",
-                "ban `rm -rf`, `sudo`, `chmod` in bash",
-                "preset: dangerous shell commands",
-            ),
-            (
-                "dangerous_sql_verbs",
-                "ban `DROP`, `TRUNCATE` in `execute_sql`",
-                "preset: dangerous SQL verbs",
-            ),
-            (
-                "irreversible_once",
-                "`deploy_production` at most once per session",
-                "irreversible action protection",
-            ),
-            (
-                "confirm_after_source",
-                "after `fetch_url`, `file_write` requires confirmation",
-                "narrow source→action gate (A,E pair)",
-            ),
-        ],
-        "cyan",
-    )
-
-    # --- Atom extensions (3) ---
-    _section(
-        "Resource & Delegation Constraints (3 det)",
-        [
-            (
-                "token_budget",
-                "session total tokens must not exceed 100000",
-                "limit token consumption",
-            ),
-            (
-                "arg_value_range",
-                "tool `set_price` field `amount` in [0, 1000]",
-                "constrain numeric arguments",
-            ),
-            (
-                "delegation_depth_limit",
-                "delegation chain max depth 3",
-                "limit agent-to-agent delegation",
-            ),
-        ],
-        "cyan",
-    )
-
-    # --- Workflow hygiene (6) ---
-    _section(
-        "Workflow Hygiene Patterns (6 det)",
-        [
-            (
-                "dry_run_before_commit",
-                "`plan_migration` dry-run before `apply_migration`",
-                "require dry-run before committing changes",
-            ),
-            (
-                "backup_before_destructive",
-                "`snapshot_db` before destructive `drop_table`",
-                "require backup before destructive action",
-            ),
-            (
-                "audit_after",
-                "`transfer_funds` must be followed by `audit_transfer`",
-                "require audit/log after sensitive action",
-            ),
-            (
-                "approval_freshness",
-                "`approve_deploy` authorizes `deploy` for 3 steps",
-                "expire old approvals after N steps",
-            ),
-            (
-                "sanitized_before_sink",
-                "`web_fetch` then `sanitize_input` before `send_email`",
-                "sanitize untrusted source before sink",
-            ),
-            (
-                "duplicate_call_limit",
-                "`search` args matching `invoice-42` at most 2 times",
-                "cap repeated same-argument calls",
-            ),
-        ],
-        "cyan",
-    )
-
-    # This build ships only deterministic patterns. Stochastic /
-    # LLM-judged evaluators (tone, relevance, generic LLM judge, ...)
-    # are an extension point with no implementation included;
-    # ``sponsio patterns`` shows det only.
-
-
-# ---------------------------------------------------------------------------
-# packs. list the contract packs that ship inside the distribution
-# ---------------------------------------------------------------------------
-
-
-@cli.command()
-def packs():
-    """List shipped contract packs with rule counts + include syntax.
-
-    Useful right after ``sponsio scan`` / ``sponsio onboard``: the
-    generated :file:`sponsio.yaml` references packs by ``include:``
-    spec, and this command prints the full inventory plus one-line
-    summaries so users can see what's been pulled in without opening
-    five YAML files.
-    """
-    # We walk the shipped contracts directory rather than hardcoding
-    # a table so new packs become visible the moment they're added.
-    from collections import Counter
-    from importlib.resources import files
-
-    import yaml as _yaml
-
-    try:
-        contracts_root = files("sponsio") / "contracts"
-    except (ModuleNotFoundError, FileNotFoundError):
-        click.echo("error: sponsio package not found on import path", err=True)
-        raise SystemExit(1) from None
-
-    rows = []  # (spec, desc_line, n_contracts, kinds_summary, needs_workspace)
-    for category_dir in sorted(contracts_root.iterdir()):
-        if not category_dir.is_dir():
-            continue
-        for pack_file in sorted(category_dir.iterdir()):
-            if not pack_file.is_file() or pack_file.suffix not in (".yaml", ".yml"):
-                continue
-            spec = f"sponsio:{category_dir.name}/{pack_file.stem}"
-            try:
-                text = pack_file.read_text(encoding="utf-8")
-                doc = _yaml.safe_load(text) or {}
-                # Header comment's first meaningful sentence gives the
-                # summary.  Fallback to "(no summary)" if the pack didn't
-                # follow the convention.
-                summary = "(no summary)"
-                for line in text.splitlines():
-                    stripped = line.lstrip("#").strip()
-                    if not stripped or stripped.startswith("="):
-                        continue
-                    if stripped.startswith("sponsio/contracts/"):
-                        continue
-                    summary = stripped
-                    break
-
-                agents = doc.get("agents") or {}
-                template = agents.get("*") or next(iter(agents.values()), {})
-                contracts = (template or {}).get("contracts") or []
-                n = len(contracts)
-
-                # Rough kind count. det patterns vs raw LTL.  OSS ships
-                # no sto pipeline; the third bucket is gone.
-                kinds = Counter()
-                for c in contracts:
-                    es = _contract_guarantee(c)
-                    if isinstance(es, dict):
-                        es_list = [es]
-                    elif isinstance(es, list):
-                        es_list = es
-                    else:
-                        es_list = []
-                    for e in es_list:
-                        if not isinstance(e, dict):
-                            continue
-                        if "ltl" in e and "pattern" not in e:
-                            kinds["raw"] += 1
-                        elif e.get("pattern"):
-                            kinds["det"] += 1
-
-                needs_ws = "<workspace>/" in text
-                rows.append((spec, summary, n, dict(kinds), needs_ws))
-            except Exception as exc:  # noqa: BLE001
-                rows.append((spec, f"(unreadable: {exc})", 0, {}, False))
-
-    click.echo()
-    click.echo(click.style("Shipped contract packs", bold=True))
-    click.echo()
-    for spec, summary, n, kinds, needs_ws in rows:
-        badge = " [needs workspace:]" if needs_ws else ""
-        click.echo(click.style(f"  {spec}{badge}", fg="cyan", bold=True))
-        k = ", ".join(f"{v} {k}" for k, v in kinds.items()) or f"{n} contracts"
-        click.echo(f"    {n} contracts ({k})")
-        click.echo(click.style(f"    {summary}", dim=True))
-        click.echo()
-    click.echo("Use in sponsio.yaml:")
-    click.echo("  agents:")
-    click.echo("    your_agent:")
-    click.echo("      include:")
-    for spec, *_ in rows:
-        click.echo(f"        - {spec}")
-
-
-# ---------------------------------------------------------------------------
-# validate
-# ---------------------------------------------------------------------------
-
-
-@cli.command()
-@click.argument("contracts", nargs=-1)
-@click.option(
-    "--config",
-    "-c",
-    "config_path",
-    type=click.Path(exists=True),
-    help="YAML config file (sponsio.yaml)",
-)
-@click.option("--agent", "-a", "agent_id", help="Agent ID to validate (with --config)")
-@click.option("--json", "as_json", is_flag=True, help="Output as JSON")
-@click.option(
-    "--traces",
-    "trace_paths",
-    multiple=True,
-    type=click.Path(exists=True),
-    help=(
-        "Replay each parsed contract against the trace file(s) or "
-        "directory.  Adds a per-contract pass/fail/error count so you "
-        "can see whether a rule would have hit your historical traffic "
-        "before flipping it to enforce mode.  Repeat for multiple paths."
-    ),
-)
-def validate(contracts, config_path, agent_id, as_json, trace_paths):
-    """Validate that contract strings parse into formal patterns.
-
-    If you pass a single existing ``.yaml`` / ``.yml`` path that looks like
-    a Sponsio project file (``agents:`` or ``version:`` + ``extractor:``),
-    it is treated as ``--config`` automatically so ``sponsio validate
-    ./sponsio.yaml`` does the right thing.
-
-    With ``--traces``, each successfully-parsed deterministic contract is
-    replayed against the supplied trace files / directories and a
-    pass / fail / error count is reported alongside the parse result.
-    Counts only. for per-failure attribution and repair suggestions
-    see the proprietary ``sponsio-pro`` validation pipeline.
-
-    Examples:\n
-        sponsio validate "tool `A` must precede `B`"\n
-        sponsio validate --config sponsio.yaml\n
-        sponsio validate --config sponsio.yaml --agent customer_bot\n
-        sponsio validate --config sponsio.yaml --traces traces/\n
-        sponsio validate ./sponsio.yaml   # same as --config when file looks like a project config
-    """
-    from sponsio.generation.dsl_to_contract import (
-        ContractSyntaxError,
-        parse_nl_unified,
-    )
-
-    if config_path and contracts:
-        click.echo(
-            click.style(
-                "Error: cannot use both --config and positional contracts", fg="red"
-            )
-        )
-        sys.exit(1)
-
-    # ``sponsio validate ./sponsio.yaml`` (forgot --config) used to try to
-    # parse the *path string* as a contract. When the path exists and the
-    # head of the file looks like a project config, treat it as --config.
-    if not config_path and len(contracts) == 1:
-        raw = contracts[0]
-        p = Path(os.path.expanduser(str(raw)))
-        if not p.is_absolute():
-            p = Path.cwd() / p
-        try:
-            p = p.resolve()
-        except OSError:
-            p = Path(raw)
-        if p.is_file() and p.suffix.lower() in (".yaml", ".yml"):
-            if _looks_like_sponsio_config(p):
-                if not as_json:
-                    click.echo(
-                        click.style("  note: ", fg="cyan", dim=True)
-                        + (
-                            f"treating {p} as a Sponsio config (equivalent to "
-                            f"`--config {p.name}`). "
-                            f"If you meant a one-line contract that looks like a path, "
-                            f"quote it or use `sponsio validate --config` explicitly."
-                        ),
-                        err=True,
-                    )
-                config_path = str(p)
-                contracts = ()
-
-    if agent_id and not config_path:
-        click.echo(click.style("Error: --agent requires --config", fg="red"))
-        sys.exit(1)
-
-    if not config_path and not contracts:
-        click.echo("Usage: sponsio validate [CONTRACTS...] or --config FILE")
-        sys.exit(1)
-
-    # ---- trace replay setup -------------------------------------------
-    # Loaded once so a 1000-contract config doesn't re-parse the trace
-    # bundle 1000 times.  ``trace_paths`` is empty in the common case.
-    traces_loaded: list = []
-    if trace_paths:
-        from sponsio.discovery.loaders import load_traces
-
-        try:
-            traces_loaded = load_traces(list(trace_paths))
-        except Exception as e:  # noqa: BLE001
-            click.echo(
-                click.style("Error: ", fg="red")
-                + f"failed to load traces from {list(trace_paths)}: {e}",
-                err=True,
-            )
-            sys.exit(1)
-        if not as_json and not traces_loaded:
-            click.echo(
-                click.style("  warn: ", fg="yellow")
-                + "no traces loaded. replay counts will all be 0",
-                err=True,
-            )
-
-    # Collect contracts to validate (flatten contract entries into
-    # per-section lists for display).
-    def _flatten(ac) -> dict:
-        assumptions: list = []
-        guarantees: list = []
-        for ce in ac.contracts:
-            if ce.assumption is not None:
-                if isinstance(ce.assumption, list):
-                    assumptions.extend(ce.assumption)
-                else:
-                    assumptions.append(ce.assumption)
-            if ce.guarantee is not None:
-                if isinstance(ce.guarantee, list):
-                    guarantees.extend(ce.guarantee)
-                else:
-                    guarantees.append(ce.guarantee)
-        return {"assumptions": assumptions, "guarantees": guarantees}
-
-    agent_contracts: dict[str, dict] = {}
-
-    if config_path:
-        from sponsio.config import load_config
-
-        config = load_config(config_path)
-        agents_to_check = (
-            {agent_id: config.agents[agent_id]} if agent_id else config.agents
-        )
-        for aid, ac in agents_to_check.items():
-            agent_contracts[aid] = _flatten(ac)
-    else:
-        agent_contracts["(inline)"] = {
-            "assumptions": [],
-            "guarantees": list(contracts),
-        }
-
-    # Validate each contract
-    all_results = []
-    all_ok = True
-
-    for aid, ag in agent_contracts.items():
-        if not as_json:
-            click.echo(click.style(f"\nAgent: {aid}", bold=True))
-
-        for section, label in [
-            ("assumptions", "Assumptions"),
-            ("guarantees", "Guarantees"),
-        ]:
-            items = ag[section]
-            if not items:
-                continue
-            if not as_json:
-                click.echo(click.style(f"  {label}:", dim=True))
-
-            for entry in items:
-                # Handle both ConstraintEntry (from config) and plain strings
-                from sponsio.config import ConstraintEntry, _compile_structured
-
-                # Track the compiled formula (or DetFormula wrapper) so
-                # the replay path below has a single source of truth
-                # regardless of which branch produced it.
-                formula_for_replay = None
-                # ``result`` is only set in the NL branches; init here
-                # so the replay-eligibility check below doesn't trip
-                # UnboundLocalError on structured / ltl entries.
-                result = None
-
-                if isinstance(entry, ConstraintEntry):
-                    if entry.is_structured:
-                        try:
-                            compiled = _compile_structured(entry)
-                            ok = True
-                            pattern = entry.pattern
-                            formula = (
-                                repr(compiled.formula)
-                                if hasattr(compiled, "formula")
-                                else ""
-                            )
-                            # OSS only ships deterministic patterns;
-                            # ``_compile_structured`` raises on unknown
-                            # names rather than falling through to sto.
-                            kind = "DET"
-                            nl = f"{entry.pattern}({', '.join(str(a) for a in entry.args)})"
-                            formula_for_replay = compiled
-                        except Exception as e:
-                            ok = False
-                            pattern = entry.pattern or ""
-                            formula = ""
-                            kind = "ERROR"
-                            nl = str(e)
-                    elif entry.is_ltl:
-                        from sponsio.config import _compile_ltl
-
-                        try:
-                            compiled = _compile_ltl(entry)
-                            ok = True
-                            pattern = "ltl"
-                            formula = repr(compiled.formula)
-                            kind = "DET"
-                            nl = entry.ltl or ""
-                            formula_for_replay = compiled
-                        except Exception as e:
-                            ok = False
-                            pattern = "ltl"
-                            formula = ""
-                            kind = "ERROR"
-                            nl = str(e)
-                    else:
-                        nl = entry.nl
-                        try:
-                            result = parse_nl_unified(nl)
-                        except ContractSyntaxError as e:
-                            ok = False
-                            pattern = ""
-                            formula = ""
-                            kind = "SYNTAX-ERROR"
-                            nl = f"{entry.nl}  ({e.hint or 'no pattern matched'})"
-                            result = None
-                        if result is None:
-                            pass  # already populated above
-                        elif result.is_det:
-                            ok = True
-                            pattern = getattr(result.hard, "pattern_name", "")
-                            formula = (
-                                repr(result.hard.formula)
-                                if hasattr(result.hard, "formula")
-                                else ""
-                            )
-                            kind = "DET"
-                            formula_for_replay = result.hard
-                else:
-                    nl = str(entry)
-                    try:
-                        result = parse_nl_unified(nl)
-                    except ContractSyntaxError as e:
-                        ok = False
-                        pattern = ""
-                        formula = ""
-                        kind = "SYNTAX-ERROR"
-                        nl = f"{str(entry)}  ({e.hint or 'no pattern matched'})"
-                        result = None
-
-                    if result is None:
-                        pass  # already populated above
-                    elif result.is_det:
-                        ok = True
-                        pattern = getattr(result.hard, "pattern_name", "")
-                        formula = (
-                            repr(result.hard.formula)
-                            if hasattr(result.hard, "formula")
-                            else ""
-                        )
-                        kind = "DET"
-                        formula_for_replay = result.hard
-                    else:
-                        pattern = ""
-                        formula = ""
-                        kind = "UNKNOWN"
-                        all_ok = False
-
-                # Replay against historical traces \u2014 only meaningful for
-                # successfully-parsed DET contracts (sto contracts need
-                # an LLM judge, which sponsio-pro covers).
-                replay_summary: dict | None = None
-                if (
-                    traces_loaded
-                    and ok
-                    and kind == "DET"
-                    and formula_for_replay is not None
-                ):
-                    from sponsio.discovery.trace_replay import replay_formula
-
-                    rep = replay_formula(formula_for_replay, traces_loaded)
-                    replay_summary = {
-                        "pass": rep.pass_count,
-                        "fail": rep.fail_count,
-                        "error": rep.error_count,
-                        "pass_rate": rep.pass_rate,
-                        "errors": list(rep.errors),
-                    }
-
-                entry = {
-                    "nl": nl,
-                    "ok": ok,
-                    "type": kind.lower(),
-                    "pattern": pattern,
-                    "formula": formula,
-                    "agent": aid,
-                    "section": section,
-                }
-                if replay_summary is not None:
-                    entry["replay"] = replay_summary
-                all_results.append(entry)
-                if not ok:
-                    all_ok = False
-
-                if not as_json:
-                    icon = (
-                        click.style("\u2713", fg="green")
-                        if ok
-                        else click.style("\u2717", fg="red")
-                    )
-                    kind_color = "cyan" if kind == "DET" else "magenta"
-                    click.echo(f"    {icon} {click.style(kind, fg=kind_color)}: {nl}")
-                    if pattern:
-                        click.echo(click.style(f"      Pattern : {pattern}", dim=True))
-                    if formula:
-                        click.echo(click.style(f"      Formula : {formula}", dim=True))
-                    if replay_summary is not None:
-                        rate = replay_summary["pass_rate"]
-                        rate_str = "n/a" if rate is None else f"{rate:.0%}"
-                        replay_line = (
-                            f"      Replay  : "
-                            f"{replay_summary['pass']} pass / "
-                            f"{replay_summary['fail']} fail"
-                        )
-                        if replay_summary["error"]:
-                            replay_line += f" / {replay_summary['error']} error"
-                        replay_line += f"  ({rate_str})"
-                        # Color: green if no fails+errors, yellow if any
-                        # fails / errors (the contract would block, or
-                        # a trace was malformed).
-                        color = (
-                            "green"
-                            if replay_summary["fail"] == 0
-                            and replay_summary["error"] == 0
-                            else "yellow"
-                        )
-                        click.echo(click.style(replay_line, fg=color, dim=True))
-
-    if as_json:
-        click.echo(json.dumps({"contracts": all_results, "ok": all_ok}, indent=2))
-    else:
-        click.echo()
-        if all_ok:
-            click.echo(
-                click.style(
-                    f"  \u2713 All {len(all_results)} contract(s) validated", fg="green"
-                )
-            )
-        else:
-            fails = sum(1 for r in all_results if not r["ok"])
-            click.echo(
-                click.style(f"  \u2717 {fails} contract(s) failed to parse", fg="red")
-            )
-        click.echo()
-
-    # Non-zero exit on any failure so CI / pre-commit hooks catch
-    # unparseable contracts instead of silently shipping them.
-    if not all_ok:
-        sys.exit(1)
-
-
-# ---------------------------------------------------------------------------
-# check
-# ---------------------------------------------------------------------------
-
-
-@cli.command()
-@click.option(
-    "--trace",
-    "-t",
-    "trace_path",
-    required=True,
-    type=click.Path(exists=True),
-    help=(
-        "Trace file to check against. Accepts OTLP/JSON, OTLP JSONL, "
-        "native Sponsio JSON/JSONL, and session JSONL. format is "
-        "sniffed from content."
-    ),
-)
-@click.argument("contracts", nargs=-1)
-@click.option(
-    "--config",
-    "-c",
-    "config_path",
-    type=click.Path(exists=True),
-    help="YAML config file (sponsio.yaml)",
-)
-@click.option("--agent", "-a", "agent_id", help="Agent ID (with --config)")
-@click.option("--json", "as_json", is_flag=True, help="Output as JSON")
-def check(trace_path, contracts, config_path, agent_id, as_json):
-    """Check contracts against an OTEL trace file.
-
-    Examples:\n
-        sponsio check --trace trace.json "tool `A` must precede `B`"\n
-        sponsio check --trace trace.json --config sponsio.yaml --agent bot
-    """
-    from sponsio.formulas.evaluator import evaluate as eval_formula
-    from sponsio.tracer.grounding import ground
-
-    if config_path and contracts:
-        click.echo(
-            click.style(
-                "Error: cannot use both --config and positional contracts", fg="red"
-            )
-        )
-        sys.exit(1)
-
-    if agent_id and not config_path:
-        click.echo(click.style("Error: --agent requires --config", fg="red"))
-        sys.exit(1)
-
-    if not config_path and not contracts:
-        click.echo("Usage: sponsio check --trace FILE [CONTRACTS...] or --config FILE")
-        sys.exit(1)
-
-    # Load trace(s) through the unified loader so this command handles
-    # the same formats as `sponsio scan --trace`.  For multi-trace
-    # files (native array, native JSONL), we concatenate events into
-    # one logical trace since `check` is a single-trace tool.
-    from sponsio.discovery.loaders import load_trace
-    from sponsio.models.trace import Trace as _Trace
-
-    try:
-        loaded = load_trace(trace_path)
-    except (FileNotFoundError, IsADirectoryError, ValueError) as e:
-        # Symmetric error handling with `sponsio scan -t`: any user-input
-        # problem surfaces as a friendly red line rather than a traceback.
-        # ``click.Path(exists=True)`` already blocks the FileNotFound case
-        # for direct args, but keeping it here protects future changes
-        # (e.g. accepting globs) from regressing.
-        click.echo(click.style(f"Error: {e}", fg="red"))
-        sys.exit(1)
-
-    if len(loaded) == 1:
-        trace = loaded[0]
-    else:
-        # Flatten. renumber ts so ordering is preserved across files.
-        merged_events: list = []
-        for t in loaded:
-            for ev in t.events:
-                merged_events.append(ev)
-        trace = _Trace(events=merged_events)
-        click.echo(
-            click.style(
-                f"  note: merged {len(loaded)} traces into one for evaluation",
-                fg="cyan",
-                dim=True,
-            ),
-            err=True,
-        )
-
-    if not trace.events:
-        click.echo(click.style("Warning: trace is empty (no spans found)", fg="yellow"))
-        sys.exit(0)
-
-    # Collect contracts (flatten ContractEntry list for this command; per-contract
-    # A->E gating is still handled in the evaluation loop below).
-    assumptions: list = []
-    guarantees: list = []
-    check_agent = agent_id or "(inline)"
-
-    if config_path:
-        from sponsio.config import load_config
-
-        config = load_config(config_path)
-        if not agent_id:
-            if len(config.agents) == 1:
-                agent_id = next(iter(config.agents))
-            else:
-                click.echo(
-                    click.style(
-                        f"Error: multiple agents in config ({list(config.agents.keys())}), "
-                        "use --agent to specify",
-                        fg="red",
-                    )
-                )
-                sys.exit(1)
-        check_agent = agent_id
-        ac = config.agents[agent_id]
-        for ce in ac.contracts:
-            if ce.assumption is not None:
-                if isinstance(ce.assumption, list):
-                    assumptions.extend(ce.assumption)
-                else:
-                    assumptions.append(ce.assumption)
-            if ce.guarantee is not None:
-                if isinstance(ce.guarantee, list):
-                    guarantees.extend(ce.guarantee)
-                else:
-                    guarantees.append(ce.guarantee)
-    else:
-        guarantees = list(contracts)
-
-    if not as_json:
-        click.echo()
-        click.echo(click.style(f"Checking: {check_agent}", bold=True))
-        click.echo(
-            click.style(f"  Trace: {trace_path} ({len(trace.events)} events)", dim=True)
-        )
-        click.echo()
-
-    # Ground the trace
-    valuations = ground(trace)
-
-    # Check assumptions
-    results = []
-    all_pass = True
-
-    if assumptions:
-        if not as_json:
-            click.echo(click.style("  Assumptions:", dim=True))
-        for entry in assumptions:
-            nl, parsed = _resolve_entry(entry)
-            if not parsed or not parsed.is_det:
-                results.append(
-                    {
-                        "nl": nl,
-                        "section": "assume",
-                        "passed": False,
-                        "note": "unparseable",
-                    }
-                )
-                all_pass = False
-                if not as_json:
-                    icon = click.style("\u2717", fg="red")
-                    click.echo(f"    {icon} {nl}  (unparseable)")
-                continue
-
-            holds = eval_formula(parsed.hard.formula, valuations)
-            results.append({"nl": nl, "section": "assume", "passed": holds})
-            if not holds:
-                all_pass = False
-            if not as_json:
-                icon = (
-                    click.style("\u2713", fg="green")
-                    if holds
-                    else click.style("\u2717", fg="red")
-                )
-                verdict = (
-                    click.style("pass", fg="green")
-                    if holds
-                    else click.style("VIOLATED", fg="red")
-                )
-                click.echo(f"    {icon} {nl} \u2014 {verdict}")
-
-    # Check guarantees
-    if guarantees:
-        if not as_json:
-            click.echo(click.style("  Guarantees:", dim=True))
-        for entry in guarantees:
-            nl, parsed = _resolve_entry(entry)
-            if not parsed or not parsed.is_det:
-                results.append(
-                    {
-                        "nl": nl,
-                        "section": "enforce",
-                        "passed": False,
-                        "note": "unparseable",
-                    }
-                )
-                all_pass = False
-                if not as_json:
-                    icon = click.style("\u2717", fg="red")
-                    click.echo(f"    {icon} {nl}  (unparseable)")
-                continue
-
-            holds = eval_formula(parsed.hard.formula, valuations)
-            results.append({"nl": nl, "section": "enforce", "passed": holds})
-            if not holds:
-                all_pass = False
-            if not as_json:
-                icon = (
-                    click.style("\u2713", fg="green")
-                    if holds
-                    else click.style("\u2717", fg="red")
-                )
-                verdict = (
-                    click.style("pass", fg="green")
-                    if holds
-                    else click.style("VIOLATED", fg="red")
-                )
-                click.echo(f"    {icon} {nl} \u2014 {verdict}")
-
-    # Summary
-    if as_json:
-        click.echo(
-            json.dumps(
-                {"agent": check_agent, "results": results, "all_pass": all_pass},
-                indent=2,
-            )
-        )
-    else:
-        click.echo()
-        total = len([r for r in results if "note" not in r])
-        passed = len([r for r in results if r["passed"] and "note" not in r])
-        if all_pass:
-            click.echo(
-                click.style(f"  \u2713 All {total} contract(s) satisfied", fg="green")
-            )
-        else:
-            fails = total - passed
-            click.echo(
-                click.style(f"  \u2717 {fails}/{total} contract(s) VIOLATED", fg="red")
-            )
-        click.echo()
-
-
-# ---------------------------------------------------------------------------
-# explain. show one contract's source + compiled form + last violation
-# ---------------------------------------------------------------------------
-
-
-@cli.command()
-@click.argument("query")
-@click.option(
-    "--config",
-    "config_path",
-    default=None,
-    type=click.Path(exists=True, dir_okay=False),
-    help="YAML config (default: ./sponsio.yaml or $SPONSIO_CONFIG).",
-)
-@click.option(
-    "--agent",
-    "agent_id",
-    default=None,
-    help="When the config has multiple agents, pick one.",
-)
-@click.option(
-    "--format",
-    "fmt",
-    type=click.Choice(["text", "json"], case_sensitive=False),
-    default="text",
-    show_default=True,
-    help="Output format.",
-)
-@click.option(
-    "--no-color",
-    is_flag=True,
-    default=False,
-    help="Disable Rich color output (text mode only).",
-)
-def explain(
-    query: str,
-    config_path: str | None,
-    agent_id: str | None,
-    fmt: str,
-    no_color: bool,
-):
-    """Explain a contract. source, compiled formula, last violation.
-
-    \b
-    Examples:
-      sponsio explain C1                       # by alias from the session view
-      sponsio explain "code freeze"            # by substring of the desc
-      sponsio explain C1 --format json         # machine-readable
-
-    The contract is resolved against the YAML config (default
-    ``./sponsio.yaml`` or ``$SPONSIO_CONFIG``). Pass ``--agent`` if the
-    config has multiple agents.
-
-    Output covers what's structurally inferable from the contract +
-    Sponsio's local session log:
-      - the assume / enforce pattern + arguments as written
-      - the compiled LTL form via ``formulas.nl_gen.formula_to_nl``
-      - the most recent BLOCKED / OBSERVED event for this contract
-        (scanning ``~/.sponsio/sessions/<agent>/*.jsonl``)
-      - generic resolution hints based on pattern shape
-
-    Richer overlays (LLM-driven contextual fix hints, cross-trace
-    pattern stats) are an extension point not part of this build.
-    """
-    import os
-
-    from sponsio.config import load_config, config_to_guard_kwargs
-    from sponsio.models.agent import Agent
-    from sponsio.models.contract import make_contracts
-    from sponsio.render.explain import (
-        explain_to_dict,
-        find_last_violation,
-        render_explain,
-        resolve_contract,
-    )
-
-    # Resolve config path: --config > $SPONSIO_CONFIG > ./sponsio.yaml.
-    cfg_path: Path | None = (
-        Path(config_path)
-        if config_path
-        else (
-            Path(os.environ["SPONSIO_CONFIG"])
-            if os.environ.get("SPONSIO_CONFIG")
-            else (Path("sponsio.yaml") if Path("sponsio.yaml").is_file() else None)
-        )
-    )
-    if cfg_path is None:
-        click.echo(
-            click.style("Error: ", fg="red")
-            + "no config found. Pass --config or create ./sponsio.yaml.",
-            err=True,
-        )
-        raise SystemExit(2)
-
-    try:
-        config = load_config(str(cfg_path))
-    except Exception as exc:
-        click.echo(click.style(f"Error loading {cfg_path}: {exc}", fg="red"), err=True)
-        raise SystemExit(2) from exc
-
-    if agent_id is None:
-        if len(config.agents) != 1:
-            click.echo(
-                click.style("Error: ", fg="red")
-                + f"config has {len(config.agents)} agents. pass --agent to disambiguate "
-                + f"(available: {', '.join(config.agents)})",
-                err=True,
-            )
-            raise SystemExit(2)
-        agent_id = next(iter(config.agents))
-    elif agent_id not in config.agents:
-        click.echo(
-            click.style("Error: ", fg="red")
-            + f"agent {agent_id!r} not in config (available: {', '.join(config.agents)})",
-            err=True,
-        )
-        raise SystemExit(2)
-
-    kw = config_to_guard_kwargs(config, agent_id)
-    contracts = make_contracts(
-        agent=Agent(id=agent_id), contracts=kw.get("contracts") or []
-    )
-
-    if not contracts:
-        click.echo(
-            click.style("Error: ", fg="red")
-            + f"no contracts compiled for agent {agent_id!r}.",
-            err=True,
-        )
-        raise SystemExit(2)
-
-    contract, idx = resolve_contract(query, contracts)
-    if contract is None:
-        # Show the catalog as a hint.
-        click.echo(
-            click.style("Error: ", fg="red")
-            + f"no contract matched {query!r}. Available:",
-            err=True,
-        )
-        for i, c in enumerate(contracts):
-            click.echo(f"  C{i + 1}  {getattr(c, 'desc', '') or '(unnamed)'}", err=True)
-        raise SystemExit(2)
-
-    last = find_last_violation(getattr(contract, "desc", "") or "")
-
-    if fmt.lower() == "json":
-        click.echo(
-            json.dumps(
-                explain_to_dict(contract, idx, last_violation=last),
-                indent=2,
-                default=str,
-            )
-        )
-        return
-
-    from rich.console import Console
-
-    console = Console(
-        file=sys.stderr,
-        soft_wrap=True,
-        highlight=False,
-        color_system=None if no_color else "auto",
-        force_terminal=False if no_color else None,
-    )
-    render_explain(
-        console=console,
-        contract=contract,
-        index=idx,
-        last_violation=last,
-        config_path=cfg_path,
-    )
-
-
-# ---------------------------------------------------------------------------
-# replay. re-render a recorded session's view from its jsonl log
-# ---------------------------------------------------------------------------
-
-
-@cli.command()
-@click.argument("session", required=False)
-@click.option(
-    "--config",
-    "config_path",
-    default=None,
-    type=click.Path(exists=True, dir_okay=False),
-    help="YAML config (for the contracts-armed table; falls back to bare table).",
-)
-@click.option(
-    "--agent",
-    "agent_id_opt",
-    default=None,
-    help="Override the agent id derived from the session log path.",
-)
-@click.option(
-    "--list",
-    "list_only",
-    is_flag=True,
-    default=False,
-    help="List available sessions and exit.",
-)
-def replay(
-    session: str | None,
-    config_path: str | None,
-    agent_id_opt: str | None,
-    list_only: bool,
-):
-    """Re-render a recorded session in the v1 mockup form.
-
-    \b
-    Examples:
-      sponsio replay sess_4f2a            # by short ID from the session view
-      sponsio replay 20260501_120000_999  # by filename stem
-      sponsio replay /path/to/log.jsonl   # by direct path
-      sponsio replay --list               # browse available sessions
-
-    Reads ``~/.sponsio/sessions/<agent>/*.jsonl`` and rebuilds the
-    AgentTurnSpan tree the live monitor would have produced, then
-    feeds it through the same renderer the session view uses.
-
-    Pass ``--config`` to also render the "contracts armed" table from
-    the YAML. without it, only contracts mentioned in the trace are
-    surfaced.
-    """
-    import os
-
-    from rich.console import Console
-
-    from sponsio.render.replay import (
-        find_session_file,
-        list_sessions,
-        load_replay,
-    )
-    from sponsio.render.session_view import render_session
-
-    console = Console(file=sys.stderr, soft_wrap=True, highlight=False)
-
-    if list_only:
-        sessions = list_sessions()
-        if not sessions:
-            click.echo("No sessions found in ~/.sponsio/sessions/.", err=True)
-            return
-        click.echo("Available sessions (most recent first):", err=True)
-        for s in sessions:
-            click.echo(
-                f"  {s['session_id']}   agent={s['agent_id']:<24} "
-                f"{s['size_bytes']:>8} bytes   {s['stem']}",
-                err=True,
-            )
-        return
-
-    if not session:
-        click.echo(
-            click.style("Error: ", fg="red")
-            + "missing SESSION arg. Try `sponsio replay --list` to browse.",
-            err=True,
-        )
-        raise SystemExit(2)
-
-    path, agent_id = find_session_file(session)
-    if path is None:
-        click.echo(
-            click.style("Error: ", fg="red")
-            + f"no session matched {session!r}. Try `sponsio replay --list`.",
-            err=True,
-        )
-        raise SystemExit(2)
-
-    turn_spans, log_agent_id = load_replay(path)
-    if not turn_spans:
-        click.echo(
-            click.style("Note: ", fg="yellow") + f"{path} has no events.",
-            err=True,
-        )
-        return
-
-    contracts: list = []
-    final_agent_id = agent_id_opt or agent_id or log_agent_id or "(unknown)"
-    cfg_path: Path | None = (
-        Path(config_path)
-        if config_path
-        else (
-            Path(os.environ["SPONSIO_CONFIG"])
-            if os.environ.get("SPONSIO_CONFIG")
-            else (Path("sponsio.yaml") if Path("sponsio.yaml").is_file() else None)
-        )
-    )
-    if cfg_path is not None:
-        try:
-            from sponsio.config import config_to_guard_kwargs, load_config
-            from sponsio.models.agent import Agent
-            from sponsio.models.contract import make_contracts
-
-            cfg = load_config(str(cfg_path))
-            cfg_agent = (
-                final_agent_id
-                if final_agent_id in cfg.agents
-                else next(iter(cfg.agents), None)
-            )
-            if cfg_agent:
-                kw = config_to_guard_kwargs(cfg, cfg_agent)
-                contracts = make_contracts(
-                    agent=Agent(id=cfg_agent),
-                    contracts=kw.get("contracts") or [],
-                )
-        except Exception as exc:
-            click.echo(
-                click.style("Warning: ", fg="yellow")
-                + f"could not load contracts from {cfg_path}: {exc}",
-                err=True,
-            )
-
-    render_session(
-        console=console,
-        agent_id=final_agent_id,
-        mode="replay",
-        contracts=contracts,
-        turn_spans=turn_spans,
-        session_id=session if session.startswith("sess_") else None,
-    )
-
-
-# ---------------------------------------------------------------------------
-# report
-# ---------------------------------------------------------------------------
-
-
-@cli.command()
-@click.option(
-    "--since",
-    default="7d",
-    show_default=True,
-    help="Time window: 'all', '30m', '24h', '7d'.",
-)
-@click.option(
-    "--agent",
-    default=None,
-    help="Filter to one agent_id. Default: every agent under ~/.sponsio/sessions.",
-)
-@click.option(
-    "--format",
-    "fmt",
-    type=click.Choice(
-        ["auto", "rich", "markdown", "md", "html", "json", "plain"],
-        case_sensitive=False,
-    ),
-    default="auto",
-    show_default=True,
-    help=(
-        "Output format. ``auto`` picks rich for an interactive terminal, "
-        "markdown for piped/CI output, or plain when NO_COLOR is set."
-    ),
-)
-@click.option(
-    "--out",
-    "-o",
-    "out_path",
-    type=click.Path(dir_okay=False, writable=True),
-    default=None,
-    help="Write report to this file. Default: stdout.",
-)
-@click.option(
-    "--save-svg",
-    "save_svg",
-    type=click.Path(dir_okay=False, writable=True),
-    default=None,
-    help="Save the rich-rendered output to an SVG file (vector, retina-safe).",
-)
-@click.option(
-    "--live",
-    is_flag=True,
-    default=False,
-    help="Watch mode: re-render every --interval seconds. Ctrl+C to exit.",
-)
-@click.option(
-    "--interval",
-    default=2.0,
-    show_default=True,
-    type=float,
-    help="Seconds between refreshes in --live mode.",
-)
-@click.option(
-    "--base-dir",
-    type=click.Path(file_okay=False),
-    default=None,
-    help="Override the session log directory (default: ~/.sponsio/sessions).",
-)
-def report(
-    since: str,
-    agent: str | None,
-    fmt: str,
-    out_path: str | None,
-    save_svg: str | None,
-    live: bool,
-    interval: float,
-    base_dir: str | None,
-):
-    """Summarize shadow-mode session logs into a shareable report.
-
-    \b
-    Examples:
-      sponsio report                                    # rich on TTY, markdown if piped
-      sponsio report --agent support_bot --since 24h    # one agent, last day
-      sponsio report --format html -o report.html       # HTML to file
-      sponsio report --format json --since all          # machine-readable dump
-      sponsio report --save-svg report.svg              # rich + SVG export
-      sponsio report --live                             # watch mode, refreshes every 2s
-
-    Reads JSONL files written by ``mode='observe'`` (shadow mode) from
-    ``~/.sponsio/sessions/<agent_id>/*.jsonl``.  Nothing is modified.
-    """
-    # Lazy imports so `sponsio --help` stays fast.
-    from pathlib import Path
-
-    from sponsio.render import pick_format
-    from sponsio.reporting import aggregate, load_events, render
-    from sponsio.reporting.reader import parse_since
-
-    # Validate --since up front so we fail fast with a readable error.
-    try:
-        parse_since(since)
-    except ValueError as e:
-        click.echo(click.style(f"Error: {e}", fg="red"))
-        raise SystemExit(2)
-
-    bd = Path(base_dir) if base_dir else None
-    resolved_fmt = pick_format(fmt)
-
-    # SVG export requires the Rich path. promote auto/markdown to rich if asked.
-    if save_svg and resolved_fmt != "rich":
-        resolved_fmt = "rich"
-
-    def _aggregate_once():
-        events = load_events(since=since, agent=agent, base_dir=bd)
-        return aggregate(events)
-
-    def _render_text(report_obj) -> str:
-        """Non-rich text output (markdown/html/json/plain)."""
-        target = "markdown" if resolved_fmt == "plain" else resolved_fmt
-        return render(report_obj, fmt=target)
-
-    def _emit_rich(report_obj) -> None:
-        """Rich path. prints directly + optionally writes SVG."""
-        from sponsio.render.rich_report import render_report, save_svg as _save_svg
-
-        console = render_report(report_obj)
-        if save_svg:
-            _save_svg(
-                console,
-                save_svg,
-                title=f"Sponsio · report --since {since}",
-            )
-            click.echo(
-                click.style("Wrote ", fg="green") + save_svg + " (SVG export)",
-                err=True,
-            )
-
-    if live:
-        if out_path is not None:
-            click.echo(
-                click.style("Error: ", fg="red")
-                + "--live cannot be combined with --out."
-            )
-            raise SystemExit(2)
-        if save_svg is not None:
-            click.echo(
-                click.style("Error: ", fg="red")
-                + "--live cannot be combined with --save-svg."
-            )
-            raise SystemExit(2)
-        import time as _time
-
-        try:
-            while True:
-                # ANSI clear-screen + home cursor; harmless on non-TTY.
-                click.echo("\x1b[2J\x1b[H", nl=False)
-                report_obj = _aggregate_once()
-                if resolved_fmt == "rich":
-                    _emit_rich(report_obj)
-                else:
-                    click.echo(_render_text(report_obj))
-                _time.sleep(max(0.25, interval))
-        except KeyboardInterrupt:
-            click.echo("\n(live mode stopped)")
-            return
-
-    report_obj = _aggregate_once()
-    if resolved_fmt == "rich":
-        _emit_rich(report_obj)
-        if out_path is not None:
-            click.echo(
-                click.style("Note: ", fg="yellow")
-                + "--out ignored with rich format; use --save-svg for export.",
-                err=True,
-            )
-        return
-
-    out = _render_text(report_obj)
-    if out_path is None:
-        click.echo(out, nl=False)
-    else:
-        with open(out_path, "w", encoding="utf-8") as f:
-            f.write(out)
-        click.echo(
-            click.style("Wrote ", fg="green")
-            + out_path
-            + f" ({len(out)} bytes, format={resolved_fmt})"
-        )
-
-
-# ---------------------------------------------------------------------------
-# serve (dashboard server stub)
-# ---------------------------------------------------------------------------
-
-
-@cli.command()
-@click.option("--host", default="127.0.0.1", help="Bind address")
-@click.option("--port", "-p", default=DASHBOARD_DEFAULT_PORT, type=int)
-@click.option("--dev", is_flag=True)
-def serve(host: str, port: int, dev: bool):
-    """Start the Sponsio dashboard server.
-
-    This build ships the contract runtime + CLI; the long-lived HTTP
-    backend that serves the web dashboard is not part of this
-    distribution. To inspect contract activity locally, use:
-
-    \b
-        sponsio host trace --follow      # live coloured stream
-        sponsio report --since 1h        # session log summary
-        sponsio replay <session>         # re-render a recorded session
-        sponsio export-sessions --to ... # ship audit to your collector
-    """
-    click.echo(
-        click.style("sponsio serve", bold=True)
-        + ": the dashboard server is not part of this distribution "
-        "(the engine ships CLI + runtime only).\n"
-        "  sponsio host trace --follow  # live alternative\n"
-        "  sponsio replay <session>     # re-render a recorded session view\n"
-        "  sponsio report --since 1h    # session-log summary\n",
-        err=True,
-    )
-    raise SystemExit(2)
-
-
-# ---------------------------------------------------------------------------
-# scan
-# ---------------------------------------------------------------------------
-
-
-@cli.command()
-@click.argument("paths", nargs=-1, required=True)
-@click.option("--agent", "-a", default="agent", help="Agent ID for generated config")
-@click.option(
-    "--llm", is_flag=True, help="Enable LLM inference (auto-detects provider from env)"
-)
-@click.option("--model", "-m", default=None, help="LLM model (default: auto-detect)")
-@click.option(
-    "--provider",
-    default=None,
-    type=click.Choice(["openai", "anthropic", "gemini"]),
-    help=(
-        "LLM provider (default: auto-detect from env). "
-        "Anthropic uses ANTHROPIC_API_KEY; Gemini uses GOOGLE_API_KEY "
-        "or GEMINI_API_KEY (1500 req/day free tier)."
-    ),
-)
-@click.option(
-    "--base-url",
-    default=None,
-    help=(
-        "OpenAI-compatible HTTP endpoint. Covers Ollama (local), "
-        "OpenRouter, DeepSeek, Together, Groq, vLLM, Azure OpenAI. "
-        "Reads OPENAI_BASE_URL env if not given."
-    ),
-)
-@click.option(
-    "--out",
-    "-o",
-    type=click.Path(),
-    default=None,
-    help=(
-        "Write YAML to this path. Defaults to `./sponsio.yaml`. "
-        "Use `-o -` to print to stdout for piping."
-    ),
-)
-@click.option(
-    "--append", is_flag=True, help="Append to existing file instead of overwriting"
-)
-@click.option(
-    "--policy",
-    "-p",
-    multiple=True,
-    type=click.Path(exists=True),
-    help="Policy document (.md/.txt) to extract constraints from",
-)
-@click.option(
-    "--push/--no-push",
-    default=False,
-    help=(
-        "Push the YAML to the local dashboard at --push-url "
-        "(default: off). The dashboard is an optional observability "
-        "companion; opt in explicitly so `sponsio scan` is a pure, "
-        "offline code-gen step by default."
-    ),
-)
-@click.option(
-    "--push-url",
-    default="http://127.0.0.1:8000",
-    help="Dashboard URL to push to (default: http://127.0.0.1:8000)",
-)
-@click.option(
-    "--config",
-    "-c",
-    "config_path",
-    type=click.Path(exists=True, dir_okay=False),
-    default=None,
-    help=(
-        "Read provider/model/api_key from sponsio.yaml's `extractor:` "
-        "section.  Implies --llm.  Explicit --provider/--model/--base-url "
-        "still win over YAML values."
-    ),
-)
-@click.option(
-    "--emit-context",
-    "emit_context",
-    is_flag=True,
-    default=False,
-    help=(
-        "Skip the LLM step and instead emit the structured inputs "
-        "(framework / tool inventory / scanned code excerpts / policy "
-        "docs) as JSON to stdout.  Used by the host "
-        "agent driving the ``sponsio`` skill: pair with "
-        "``sponsio prompt scan`` and apply in the agent's own LLM "
-        "context. no UnifiedExtractor call, no extra API key."
-    ),
-)
-def scan(
-    paths: tuple[str, ...],
-    agent: str,
-    llm: bool,
-    model: str | None,
-    provider: str | None,
-    base_url: str | None,
-    out: str | None,
-    append: bool,
-    policy: tuple[str, ...],
-    push: bool,
-    push_url: str,
-    config_path: str | None,
-    emit_context: bool,
-):
-    """Scan source code and policy docs to propose contracts.
-
-    For first-time setup, prefer ``sponsio onboard``. it composes
-    framework detection + scan + ``init``-style provider config +
-    ``doctor`` health checks into a single command.  ``scan`` is the
-    library-maintenance tool you reach for *after* you have a
-    ``sponsio.yaml``: re-mine contracts from new code or append from
-    a policy doc.
-
-    Analyzes tool definitions, decorators, and call patterns to infer
-    safety constraints. Optionally extracts constraints from policy
-    documents (.md/.txt) using the discovered tool inventory as context.
-
-    \b
-    Examples:
-      sponsio scan src/                                # writes ./sponsio.yaml (rule-based)
-      sponsio scan src/ --llm                          # + LLM inference
-      sponsio scan src/ --policy security.md --llm     # code + policy
-      sponsio scan src/ -o custom.yaml                 # write to custom path
-      sponsio scan src/ -o sponsio.yaml --append       # merge into existing
-      sponsio scan src/ -o -                           # print to stdout (pipe)
-      sponsio scan src/ --push                         # also push to dashboard
-    """
-    from sponsio.discovery.extractors.code_analysis import CodeAnalyzer
-
-    # Route progress messages to stderr with light styling so the YAML
-    # body on stdout is still pipeable to a file or another command.
-    def _scan_progress(msg: str) -> None:
-        if emit_context:
-            return
-        click.echo(click.style("· ", fg="cyan", dim=True) + msg, err=True)
-
-    # ---- agent-driven path: dump inputs, skip LLM step ------------------
-    # ``--emit-context`` runs the deterministic scan stages (AST tool
-    # inventory, policy doc collection) and stops short of the LLM
-    # contract-mining inside ``CodeAnalyzer.generate_yaml``.
-    # The host agent picks up using ``sponsio prompt scan``.
-    if emit_context:
-        analyzer = CodeAnalyzer(use_llm=False)
-        source_paths = list(paths)
-        tool_inventory = analyzer.get_tool_inventory(source_paths) or []
-
-        policy_docs: list[dict] = []
-        for p in policy:
-            try:
-                policy_docs.append(
-                    {
-                        "path": str(p),
-                        "content": Path(p).read_text(encoding="utf-8"),
-                    }
-                )
-            except OSError:
-                continue
-
-        existing_yaml_text = ""
-        out_path = Path(out) if out and out != "-" else Path("sponsio.yaml")
-        if out_path.exists():
-            try:
-                existing_yaml_text = out_path.read_text(encoding="utf-8")
-            except OSError:
-                pass
-
-        click.echo(
-            json.dumps(
-                {
-                    "agent_id": agent,
-                    "source_paths": source_paths,
-                    "tool_inventory": tool_inventory,
-                    "policy_docs": policy_docs,
-                    "existing_yaml": existing_yaml_text,
-                    "out_path": str(out_path),
-                    "next_steps_hint": (
-                        "Run ``sponsio prompt scan`` to get the prompt "
-                        "template, apply it to this JSON in your own LLM "
-                        f"context, then write the resulting YAML to {out_path} "
-                        "via Edit/Write.  Validate with "
-                        f"``sponsio validate --config {out_path}``."
-                    ),
-                },
-                indent=2,
-                ensure_ascii=False,
-            )
-        )
-        return
-
-    # Pull provider/model/key/base_url from the YAML's ``extractor:``
-    # section if --config was given.  CLI flags retain the highest
-    # precedence. they're how you override on a one-off basis.
-    api_key: str | None = None
-    if config_path:
-        from sponsio.config import load_config
-
-        cfg = load_config(config_path)
-        ext = cfg.extractor
-        if not (ext.provider or ext.model or ext.api_key or ext.base_url):
-            click.echo(
-                click.style("  warn: ", fg="yellow")
-                + f"{config_path} has no `extractor:` section. "
-                "nothing to inherit.",
-                err=True,
-            )
-        else:
-            _scan_progress(
-                f"using extractor config from {config_path} "
-                f"(provider={ext.provider or '<auto>'}, "
-                f"model={ext.model or '<default>'})"
-            )
-        provider = provider or ext.provider
-        model = model or ext.model
-        base_url = base_url or ext.base_url
-        api_key = ext.api_key
-        # --config implies --llm: configuring an extractor and then NOT
-        # using it would be confusing.
-        if not llm:
-            llm = True
-            _scan_progress("--config implies --llm; enabling LLM inference")
-
-    analyzer = CodeAnalyzer(
-        use_llm=llm,
-        llm_model=model,
-        api_key=api_key,
-        provider=provider,
-        base_url=base_url,
-        progress=_scan_progress,
-    )
-    source_paths = list(paths)
-
-    # Extract tool inventory for policy document context
-    tool_inventory = analyzer.get_tool_inventory(source_paths) if policy else None
-
-    yaml_content = analyzer.generate_yaml(
-        source_paths,
-        agent_id=agent,
-        policy_paths=list(policy),
-        tool_inventory=tool_inventory,
-    )
-
-    # --- Auto-validate & drop unparseable contracts ---------------------
-    # Goal: the file we hand the user is *directly usable*. Any contract
-    # that the parser can't compile is dropped here (and listed on
-    # stderr) instead of being left as a landmine in the YAML.
-    yaml_content, dropped_contracts = _filter_invalid_contracts(yaml_content)
-
-    # --- Post-scan summary (stderr) -------------------------------------
-    # Helps users notice the "0 contracts" case immediately and points
-    # them at --llm if they only ran the AST pass.
-    n_tools, n_contracts, n_review = _scan_summary_counts(yaml_content)
-    summary_color = "green" if n_contracts > 0 else "yellow"
-    summary = f"Scan summary: {n_tools} tool(s), {n_contracts} contract(s) kept"
-    if n_review:
-        summary += f" ({n_review} flagged for review)"
-    if dropped_contracts:
-        summary += f", {len(dropped_contracts)} dropped (failed to parse)"
-    click.echo(click.style("• " + summary, fg=summary_color), err=True)
-    for d in dropped_contracts:
-        click.echo(
-            click.style("  dropped: ", fg="yellow")
-            + f"[{d['agent']}] "
-            + click.style(d["nl"][:120], dim=True)
-            + (f"  ({d['error']})" if d.get("error") else ""),
-            err=True,
-        )
-    if n_tools == 0:
-        click.echo(
-            click.style("  note: ", fg="cyan")
-            + "0 tools usually means nothing in the scanned path matched "
-            "Sponsio's discovery rules (``@tool``, ``Agent(tools=[...])``, "
-            "``TOOLS = [fn, ...]``, etc.), or the tree was effectively empty "
-            "(dependency dirs like ``.venv`` / ``node_modules`` are skipped). "
-            "Point at the directory that contains your agent's tool modules.",
-            err=True,
-        )
-    if n_contracts == 0 and not llm:
-        click.echo(
-            click.style("  hint: ", fg="cyan")
-            + "no contracts inferred from AST. Re-run with "
-            + click.style("--llm", bold=True)
-            + " (and optionally --policy <doc>) for richer inference.",
-            err=True,
-        )
-    if policy and not llm:
-        click.echo(
-            click.style("  warn: ", fg="yellow")
-            + "--policy was given but --llm was not. "
-            + f"{len(policy)} policy file(s) were ignored.",
-            err=True,
-        )
-
-    # Default output: write to ``./sponsio.yaml`` so the common
-    # interactive case never leaves the user wondering where the YAML
-    # went.  Two opt-outs:
-    #   * ``-o -``      → print to stdout (pipeline use)
-    #   * ``-o <path>`` → write to a specific path
-    if out == "-":
-        click.echo(yaml_content)
-        click.echo(
-            click.style("• ", fg="cyan")
-            + "YAML written to stdout (use `-o <path>` to save to a file).",
-            err=True,
-        )
-    else:
-        target = out or "sponsio.yaml"
-        existed = os.path.exists(target)
-        if append and existed:
-            with open(target) as f:
-                existing = f.read()
-            yaml_content = _merge_yaml(existing, yaml_content)
-        with open(target, "w") as f:
-            f.write(yaml_content)
-        abs_out = os.path.abspath(target)
-        verb = (
-            "Updated" if append and existed else ("Overwrote" if existed else "Wrote")
-        )
-        click.echo(
-            click.style("✓ ", fg="green") + f"{verb} {click.style(abs_out, bold=True)}",
-            err=True,
-        )
-        if existed and not append:
-            click.echo(
-                click.style("  note: ", fg="yellow")
-                + "existing file was overwritten. "
-                + "Use --append to merge new contracts into it instead.",
-                err=True,
-            )
-        click.echo(
-            click.style("  tip: ", fg="cyan", dim=True)
-            + f"re-run `sponsio validate --config {abs_out}` after manual edits.",
-            err=True,
-        )
-
-    if push:
-        _push_scan_to_dashboard(
-            yaml_content=yaml_content,
-            filename=(os.path.basename(out) if out and out != "-" else "sponsio.yaml"),
-            dashboard_url=push_url,
-            source_paths=source_paths,
-        )
-
-
-def _filter_invalid_contracts(yaml_content: str) -> tuple[str, list[dict]]:
-    """Drop contracts that fail to compile so the saved YAML is usable as-is.
-
-    Walks every ``agents.<id>.contracts[*]`` entry, runs the same parser
-    that ``sponsio validate`` uses, and rewrites the YAML with only the
-    entries that parse cleanly. Bad ones are returned for stderr display.
-
-    Conservative on errors: if PyYAML / the parser modules aren't
-    importable, returns the input unchanged (and an empty drop list) so
-    a minimal install still gets a working scan, just without the
-    auto-validate net.
-
-    Returns:
-        (cleaned_yaml, dropped) where ``dropped`` is a list of
-        ``{"agent": str, "nl": str, "error": str}``.
-    """
-    try:
-        import yaml as _yaml
-    except ImportError:
-        return yaml_content, []
-
-    try:
-        from sponsio.config import (
-            _compile_ltl,
-            _compile_structured,
-            _parse_constraint_entry,
-        )
-        from sponsio.generation.dsl_to_contract import (
-            ContractSyntaxError,
-            parse_nl_unified,
-        )
-    except ImportError:
-        return yaml_content, []
-
-    try:
-        data = _yaml.safe_load(yaml_content)
-    except _yaml.YAMLError:
-        return yaml_content, []
-
-    if not isinstance(data, dict):
-        return yaml_content, []
-
-    agents_raw = data.get("agents", {})
-    if not isinstance(agents_raw, dict):
-        return yaml_content, []
-
-    def _validate_one(item) -> tuple[bool, str, str]:
-        try:
-            entry = _parse_constraint_entry(item)
-        except Exception as e:  # noqa: BLE001
-            return False, str(item)[:120], f"parse: {e}"
-        if entry.is_structured:
-            try:
-                _compile_structured(entry)
-            except Exception as e:  # noqa: BLE001
-                args = ", ".join(str(a) for a in (entry.args or []))
-                return False, f"{entry.pattern}({args})", str(e)
-            return True, "", ""
-        elif entry.is_ltl:
-            try:
-                _compile_ltl(entry)
-            except Exception as e:  # noqa: BLE001
-                return False, (entry.ltl or "")[:120], str(e)
-            return True, "", ""
-        else:
-            nl = entry.nl or ""
-            try:
-                parse_nl_unified(nl)
-            except ContractSyntaxError as e:
-                return False, nl, e.hint or "no pattern matched"
-            except Exception as e:  # noqa: BLE001
-                return False, nl, str(e)
-            return True, "", ""
-
-    bad_per_agent: dict[str, set[int]] = {}
-    dropped: list[dict] = []
-
-    for agent_id, ag in agents_raw.items():
-        # An agent block is normally a dict with `contracts:`; bare lists
-        # are tolerated by the loader but rare from generate_yaml. Handle
-        # both for safety.
-        if isinstance(ag, dict):
-            contracts = ag.get("contracts", [])
-        elif isinstance(ag, list):
-            contracts = ag
-        else:
-            continue
-        if not isinstance(contracts, list):
-            continue
-
-        bad: set[int] = set()
-        for idx, ce in enumerate(contracts):
-            # An entry can be either a bare string (E only, NL form), or
-            # a dict with A/G keys whose values are themselves NL strings
-            # or structured ``{pattern, args}`` dicts.
-            sub_items: list = []
-            if isinstance(ce, str):
-                sub_items.append(ce)
-            elif isinstance(ce, dict):
-                for key in ("A", "G"):
-                    if key not in ce:
-                        continue
-                    val = ce[key]
-                    sub_items.extend(val if isinstance(val, list) else [val])
-            else:
-                continue
-
-            entry_dropped = False
-            for it in sub_items:
-                if it is None:
-                    continue
-                ok, nl_repr, err = _validate_one(it)
-                if not ok:
-                    dropped.append(
-                        {"agent": str(agent_id), "nl": nl_repr, "error": err}
-                    )
-                    entry_dropped = True
-                    break
-            if entry_dropped:
-                bad.add(idx)
-
-        if bad:
-            bad_per_agent[str(agent_id)] = bad
-
-    if not bad_per_agent:
-        return yaml_content, dropped  # nothing to rewrite
-
-    cleaned = _drop_contract_indices(yaml_content, bad_per_agent)
-    return cleaned, dropped
-
-
-def _drop_contract_indices(
-    yaml_content: str, bad_per_agent: dict[str, set[int]]
-) -> str:
-    """Remove specific contract entries (by 0-based index) per agent.
-
-    Preserves comments, confidence tags and the surrounding YAML
-    structure that ``generate_yaml`` produces.  If an agent's
-    ``contracts:`` list ends up empty we replace it with ``contracts: []``
-    so the resulting file still parses.
-    """
-    out: list[str] = []
-    lines = yaml_content.split("\n")
-
-    in_agents = False
-    current_agent: str | None = None
-    in_contracts = False
-    current_idx = -1
-    skipping = False
-    contracts_line_idx: int | None = None
-    kept_in_current_contracts = 0
-
-    def _finalize_contracts_block() -> None:
-        # If the contracts: list ended up empty, swap the header line for
-        # ``contracts: []`` so the YAML stays valid.
-        nonlocal contracts_line_idx, kept_in_current_contracts
-        if contracts_line_idx is not None and kept_in_current_contracts == 0:
-            header = out[contracts_line_idx]
-            stripped = header.lstrip()
-            indent = header[: len(header) - len(stripped)]
-            if stripped.rstrip().endswith(":"):
-                out[contracts_line_idx] = f"{indent}contracts: []"
-        contracts_line_idx = None
-        kept_in_current_contracts = 0
-
-    for line in lines:
-        stripped = line.lstrip()
-        indent = len(line) - len(stripped)
-
-        # Blank / comment lines: keep unless we're inside a dropped entry.
-        if not stripped or stripped.startswith("#"):
-            if skipping:
-                continue
-            out.append(line)
-            continue
-
-        # Top-level key (col 0) → reset everything.
-        if indent == 0:
-            _finalize_contracts_block()
-            in_agents = stripped.startswith("agents:")
-            current_agent = None
-            in_contracts = False
-            current_idx = -1
-            skipping = False
-            out.append(line)
-            continue
-
-        # Inside agents: each agent header sits at indent 2.
-        if in_agents and indent == 2 and stripped.rstrip().endswith(":"):
-            _finalize_contracts_block()
-            current_agent = stripped.rstrip()[:-1].strip()
-            in_contracts = False
-            current_idx = -1
-            skipping = False
-            out.append(line)
-            continue
-
-        # Properties of the current agent live at indent 4.
-        if current_agent is not None and indent == 4:
-            _finalize_contracts_block()
-            in_contracts = stripped.startswith("contracts:")
-            skipping = False
-            current_idx = -1
-            if in_contracts:
-                contracts_line_idx = len(out)
-                kept_in_current_contracts = 0
-            out.append(line)
-            continue
-
-        # Inside a contracts: list, entries start at indent 6 with "- ".
-        if in_contracts and indent >= 6:
-            if stripped.startswith("- "):
-                current_idx += 1
-                bad_set = bad_per_agent.get(current_agent or "", set())
-                skipping = current_idx in bad_set
-                if not skipping:
-                    kept_in_current_contracts += 1
-                    out.append(line)
-                continue
-            # Continuation line of the current entry.
-            if not skipping:
-                out.append(line)
-            continue
-
-        # Anything else: outside our tracked regions.
-        skipping = False
-        out.append(line)
-
-    _finalize_contracts_block()
-    return "\n".join(out)
-
-
-def _scan_summary_counts(yaml_content: str) -> tuple[int, int, int]:
-    """Count tools, contracts and review-flagged contracts in scan YAML.
-
-    Tolerant to formatting; we just look for stable line shapes that the
-    YAML emitter produces.  Returns ``(tools, contracts, review_flagged)``.
-    """
-    n_tools = 0
-    n_contracts = 0
-    n_review = 0
-    in_tools = False
-    for raw in yaml_content.splitlines():
-        line = raw.rstrip()
-        stripped = line.strip()
-        if line.startswith("tools:"):
-            in_tools = True
-            continue
-        if in_tools:
-            if line.startswith("  - name:"):
-                n_tools += 1
-                continue
-            if line and not line.startswith(" "):
-                in_tools = False
-        if stripped.startswith("- G:") or stripped.startswith("- A:"):
-            n_contracts += 1
-            if "review recommended" in stripped:
-                n_review += 1
-    return n_tools, n_contracts, n_review
-
-
-def _push_scan_to_dashboard(
-    yaml_content: str,
-    filename: str,
-    dashboard_url: str,
-    source_paths: list[str],
-) -> None:
-    """POST the scan YAML to the running dashboard.
-
-    Silently skips if the dashboard isn't reachable; this is additive UX,
-    not a required step.
-    """
-    base = dashboard_url.rstrip("/")
-    try:
-        import httpx
-    except ImportError:
-        click.echo(
-            click.style("  note: ", fg="yellow")
-            + "httpx not installed, skipping dashboard push."
-        )
-        return
-
-    # 1. Check that the dashboard is actually running before uploading.
-    try:
-        r = httpx.get(f"{base}/api/health", timeout=1.5)
-        if r.status_code != 200:
-            raise RuntimeError(f"/api/health returned {r.status_code}")
-    except Exception:
-        click.echo(
-            click.style("  tip: ", fg="cyan")
-            + f"dashboard not running at {base}. start it with "
-            + click.style("sponsio serve", bold=True)
-            + " to see scan results in the UI."
-        )
-        return
-
-    # 2. POST the YAML as a file upload, tagged with source=cli so the
-    #    dashboard's CLI tab can distinguish it from browser uploads.
-    try:
-        files = {"file": (filename, yaml_content.encode("utf-8"), "text/yaml")}
-        r = httpx.post(
-            f"{base}/api/scan/upload",
-            files=files,
-            params={"source": "cli"},
-            timeout=10.0,
-        )
-        if r.status_code != 200:
-            click.echo(
-                click.style("  push failed: ", fg="yellow")
-                + f"HTTP {r.status_code} {r.text[:200]}"
-            )
-            return
-        result = r.json()
-        summary = (
-            f"{result.get('agent_name', '?')}: "
-            f"{result.get('score', 0)}/100 "
-            f"({result.get('grade', '?')})"
-        )
-        click.echo(click.style("✓ ", fg="green") + f"Pushed to dashboard. {summary}")
-        click.echo(
-            f"  View at {click.style(base.replace(':8000', ':3000') + '/scan', bold=True)}"
-        )
-    except Exception as e:
-        click.echo(click.style("  push failed: ", fg="yellow") + str(e))
-
-
-def _merge_yaml(existing: str, new: str) -> str:
-    """Merge new scan results into an existing YAML file.
-
-    Appends new contract entries (``- E:`` / ``- A: ... E:``) from
-    *new* after the last contract in *existing*, avoiding duplicates.
-
-    Works with the current ``contracts: [{A, E}]`` YAML schema.
-    """
-    existing_lines = existing.rstrip().split("\n")
-
-    # --- Extract contract entries from new content ---
-    # A contract entry starts with a line matching `- E:` or `- A:` at
-    # the expected indent (6 spaces inside `contracts:`). Continuation
-    # lines are indented deeper.
-    new_lines = new.split("\n")
-    new_entries: list[list[str]] = []
-    in_contracts = False
-    current_entry: list[str] = []
-
-    for line in new_lines:
-        stripped = line.strip()
-        if "contracts:" in line and stripped != "contracts: []":
-            in_contracts = True
-            continue
-        if not in_contracts:
-            continue
-        # A new entry starts with `- E:` or `- A:` (possibly with trailing comment)
-        if stripped.startswith("- G:") or stripped.startswith("- A:"):
-            if current_entry:
-                new_entries.append(current_entry)
-            current_entry = [line]
-        elif current_entry and (
-            stripped.startswith("pattern:")
-            or stripped.startswith("args:")
-            or stripped.startswith("source:")
-            or stripped.startswith("G:")
-            or stripped.startswith("desc:")
-        ):
-            # Continuation of the current entry
-            current_entry.append(line)
-        elif current_entry and not stripped and not stripped.startswith("#"):
-            # Blank line or end of section
-            pass
-        elif current_entry and stripped.startswith("#"):
-            # Comment inside an entry. keep it
-            current_entry.append(line)
-        elif not stripped:
-            continue
-        else:
-            # Non-entry, non-continuation line. we've left the contracts block
-            break
-    if current_entry:
-        new_entries.append(current_entry)
-
-    if not new_entries:
-        return existing
-
-    # --- Fingerprint existing entries to deduplicate ---
-    # Normalize each entry to a single key string for comparison.
-    def _fingerprint(lines: list[str]) -> str:
-        return " ".join(ln.strip() for ln in lines)
-
-    existing_fingerprints: set[str] = set()
-    temp_entry: list[str] = []
-    in_existing_contracts = False
-    for line in existing_lines:
-        stripped = line.strip()
-        if "contracts:" in line:
-            in_existing_contracts = True
-            continue
-        if not in_existing_contracts:
-            continue
-        if stripped.startswith("- G:") or stripped.startswith("- A:"):
-            if temp_entry:
-                existing_fingerprints.add(_fingerprint(temp_entry))
-            temp_entry = [line]
-        elif temp_entry and stripped and not stripped.startswith("#"):
-            temp_entry.append(line)
-        elif not stripped:
-            continue
-    if temp_entry:
-        existing_fingerprints.add(_fingerprint(temp_entry))
-
-    # Filter out duplicates
-    to_add = [
-        entry
-        for entry in new_entries
-        if _fingerprint(entry) not in existing_fingerprints
-    ]
-
-    if not to_add:
-        return existing
-
-    # --- Append after last content line ---
-    result = existing.rstrip() + "\n"
-    result += "      # --- appended by sponsio scan ---\n"
-    for entry in to_add:
-        result += "\n".join(entry) + "\n"
-    return result
-
-
-# ---------------------------------------------------------------------------
-# doctor
-# ---------------------------------------------------------------------------
-
-
-@cli.command(name="export")
-@click.argument(
-    "source",
-    type=click.Path(exists=True, file_okay=True, dir_okay=True, path_type=Path),
-)
-@click.option(
-    "--to",
-    "target_dir",
-    type=click.Path(file_okay=False, dir_okay=True, path_type=Path),
-    required=True,
-    help="Output directory for OTLP-JSON trace files.",
-)
-@click.option(
-    "--label",
-    type=click.Choice(["safe", "unsafe", "none"]),
-    default="safe",
-    show_default=True,
-    help=(
-        "Filename prefix applied to each output trace.  ``safe`` / "
-        "``unsafe`` make the file ready for `sponsio eval`; ``none`` "
-        "preserves the input basename untouched (useful when you've "
-        "already pre-labelled Sponsio-native dumps)."
-    ),
-)
-@click.option(
-    "--agent",
-    "agent_id",
-    default=None,
-    help=(
-        "Override the ``service.name`` stamped on the OTLP output.  "
-        "Defaults to the ``metadata.agent_id`` in the source JSON, "
-        "then to the first event's ``agent``, then to ``'agent'``."
-    ),
-)
-@click.option(
-    "--glob",
-    "glob_pattern",
-    default="*.json",
-    show_default=True,
-    help="Only convert files matching this glob (directory mode only).",
-)
-def export_cmd(
-    source: Path,
-    target_dir: Path,
-    label: str,
-    agent_id: str | None,
-    glob_pattern: str,
-):
-    """Convert Sponsio-native trace dumps to OTLP JSON for ``sponsio eval``.
-
-    The canonical flow from prod to eval corpus:
-
-    \b
-        # 1. In your agent (observe mode. never blocks):
-        guard = BaseGuard(agent_id="bot", contracts=[...], mode="observe")
-        # ...runs happen, violations logged but not enforced...
-
-        # 2. Dump the accumulated trace to disk at session end:
-        guard.trace.export("/var/log/sponsio/run.json")
-
-        # 3. Later, convert a directory of these dumps into an eval corpus:
-        sponsio export /var/log/sponsio/ --to traces/ --label safe
-
-        # 4. Re-label incident traces and re-run eval:
-        mv traces/safe_run_123.json traces/unsafe_run_123.json
-        sponsio eval traces/ --config sponsio.yaml
-
-    SOURCE may be a single ``.json`` file or a directory of them.
-    Output filenames are ``{label}_{source-basename}.json``. the
-    prefix is what ``sponsio eval`` reads to know which traces are
-    expected to pass vs be blocked, so picking the right ``--label``
-    at export time saves a rename pass later.
-    """
-    from sponsio.models.trace import Trace
-    from sponsio.tracer.otel_writer import trace_to_otlp
-
-    # Collect source files
-    if source.is_file():
-        sources = [source]
-    else:
-        sources = sorted(source.glob(glob_pattern))
-        if not sources:
-            click.echo(
-                click.style(
-                    f"No files matched {glob_pattern} under {source}", fg="yellow"
-                ),
-                err=True,
-            )
-            sys.exit(0)
-
-    target_dir.mkdir(parents=True, exist_ok=True)
-
-    converted = 0
-    skipped: list[tuple[Path, str]] = []
-
-    for src in sources:
-        try:
-            raw = json.loads(src.read_text())
-        except (json.JSONDecodeError, OSError) as e:
-            skipped.append((src, f"read: {e}"))
-            continue
-
-        # Accept either the bare Trace dict shape ({"events": [...], "metadata": {...}})
-        # OR the richer ``export_trace()`` envelope (same shape, extra metadata).
-        # Reject OTLP input. that's already in the target shape and would
-        # silently duplicate rather than convert.
-        if "resourceSpans" in raw:
-            skipped.append((src, "already OTLP JSON. refusing to re-wrap"))
-            continue
-        if "events" not in raw:
-            skipped.append((src, "no 'events' key. not a Sponsio trace dump"))
-            continue
-
-        try:
-            trace = Trace.from_dict(raw)
-        except (KeyError, TypeError) as e:
-            skipped.append((src, f"parse: {e}"))
-            continue
-
-        effective_agent = (
-            agent_id or (raw.get("metadata") or {}).get("agent_id") or None
-        )
-        otlp = trace_to_otlp(trace, agent_id=effective_agent)
-
-        # Figure out output filename + label prefix
-        stem = src.stem
-        if label == "none":
-            out_name = f"{stem}.json"
-        else:
-            # Don't double-prefix if the source already has safe_/unsafe_
-            lowered = stem.lower()
-            if lowered.startswith(("safe_", "safe-", "unsafe_", "unsafe-")):
-                out_name = f"{stem}.json"
-            else:
-                out_name = f"{label}_{stem}.json"
-
-        out_path = target_dir / out_name
-        out_path.write_text(json.dumps(otlp, indent=2))
-        converted += 1
-
-    click.echo(
-        click.style("✓ ", fg="green")
-        + f"Converted {converted} trace(s) to {target_dir}"
-    )
-    if skipped:
-        click.echo(click.style("  skipped:", fg="yellow"))
-        for p, why in skipped:
-            click.echo(f"    · {p.name}. {why}")
-
-
-# ---------------------------------------------------------------------------
-# `sponsio export-sessions`. push session audit log to OTLP
-# ---------------------------------------------------------------------------
-
-
-def _session_event_to_otlp_span(event: dict) -> dict:
-    """Convert one ``MonitorEvent``-shaped JSONL record into an OTLP span.
-
-    The session log captures *flat* monitor events (one row per
-    contract verdict), not the full span tree. We synthesise a
-    self-contained OTLP span per event so the dashboard's "Today's
-    blocks" card has the same attribute keys it gets from live
-    span-tree exports.
-
-    Lossy on purpose: we don't re-derive the contract_check tree from
-    flat events, so the violation card works but the rule-fire-heatmap
-    won't have per-phase precondition / guarantee detail. That's
-    acceptable for historical replay; live exports keep the full tree.
-    """
-    from sponsio.tracer import semconv
-
-    ts_unix = float(event.get("ts") or 0.0)
-    ts_ns = int(ts_unix * 1_000_000_000) if ts_unix else 0
-    result = event.get("result") or {}
-    action = result.get("action") or "allowed"
-    blocked = action in ("blocked", "escalated", "observed")
-
-    attrs: list[dict] = []
-    if event.get("agent_id"):
-        attrs.append(_attr_for_session(semconv.ATTR_AGENT_ID, event["agent_id"]))
-    if event.get("action"):
-        attrs.append(_attr_for_session(semconv.ATTR_EVENT_TOOL, event["action"]))
-    if ts_ns:
-        attrs.append(_attr_for_session(semconv.ATTR_EVENT_TIMESTAMP_NS, ts_ns))
-    if event.get("pipeline"):
-        # ``hard`` is the legacy alias; emit the public ``det`` name.
-        pipeline = "det" if event["pipeline"] == "hard" else event["pipeline"]
-        attrs.append(_attr_for_session(semconv.ATTR_CONTRACT_PIPELINE, pipeline))
-    if event.get("constraint"):
-        attrs.append(
-            _attr_for_session(semconv.ATTR_CONTRACT_LABEL, event["constraint"])
-        )
-    attrs.append(_attr_for_session(semconv.ATTR_OUTCOME_BLOCKED, bool(blocked)))
-    attrs.append(
-        _attr_for_session(
-            semconv.ATTR_OUTCOME_STATUS,
-            "violated" if blocked else "ok",
-        )
-    )
-    attrs.append(_attr_for_session(semconv.ATTR_ENFORCEMENT_ACTION, action))
-    if result.get("message"):
-        attrs.append(
-            _attr_for_session(semconv.ATTR_VIOLATION_EVIDENCE, result["message"])
-        )
-
-    return {
-        "traceId": "0" * 32,
-        "spanId": f"{int(ts_unix * 1000):016x}" if ts_ns else "0" * 16,
-        "name": semconv.SPAN_AGENT_TURN,
-        "startTimeUnixNano": str(ts_ns or 0),
-        "endTimeUnixNano": str(ts_ns or 0),
-        "status": {"code": 2 if blocked else 1},
-        "attributes": attrs,
-    }
-
-
-def _attr_for_session(key: str, value):
-    """Local copy of otel_writer._attr. used by the session importer
-    so we don't leak the writer's private API into this CLI command."""
-    if isinstance(value, bool):
-        v: dict = {"boolValue": value}
-    elif isinstance(value, int):
-        v = {"intValue": str(value)}
-    elif isinstance(value, float):
-        v = {"doubleValue": value}
-    else:
-        v = {"stringValue": str(value)}
-    return {"key": key, "value": v}
-
-
-@cli.command(name="export-sessions")
-@click.option(
-    "--since",
-    default="24h",
-    show_default=True,
-    help=(
-        "Time window relative to now: ``24h`` / ``7d`` / ``30m`` / "
-        "``90s``, or ``all`` for no cutoff. Bare numbers default to "
-        "hours."
-    ),
-)
-@click.option(
-    "--agent",
-    "agent_filter",
-    default=None,
-    help=(
-        "Only export sessions for this agent_id. Defaults to all "
-        "agents under ``~/.sponsio/sessions/``."
-    ),
-)
-@click.option(
-    "--sessions-dir",
-    "sessions_dir",
-    type=click.Path(file_okay=False, dir_okay=True, path_type=Path),
-    default=None,
-    help=(
-        "Override the source directory. Default: "
-        "``$SPONSIO_SESSIONS_DIR`` or ``~/.sponsio/sessions/``."
-    ),
-)
-@click.option(
-    "--to",
-    "destination",
-    required=True,
-    help=(
-        "Output destination. Either an OTLP file path "
-        "(``./traces.jsonl``) or an HTTP endpoint "
-        "(``https://collector.example.com/v1/traces``)."
-    ),
-)
-@click.option(
-    "--header",
-    "headers_raw",
-    multiple=True,
-    help=(
-        "Extra HTTP headers as ``Key: Value``. May be specified "
-        "multiple times. Auth keys, tenant ids etc. go here. Only "
-        "honored when ``--to`` is an HTTP URL."
-    ),
-)
-@click.option(
-    "--batch-size",
-    type=int,
-    default=50,
-    show_default=True,
-    help="Spans per HTTP POST (HTTP destination only).",
-)
-@click.option(
-    "--service-name",
-    default=None,
-    help=(
-        "OTLP ``resource.service.name`` stamped on every exported "
-        "span. Defaults to the per-agent_id of each session file."
-    ),
-)
-def export_sessions_cmd(
-    since: str,
-    agent_filter: str | None,
-    sessions_dir: Path | None,
-    destination: str,
-    headers_raw: tuple[str, ...],
-    batch_size: int,
-    service_name: str | None,
-):
-    """Ship audit-log session events to an OTLP destination.
-
-    Reads ``~/.sponsio/sessions/<agent_id>/*.jsonl``, converts each
-    ``MonitorEvent`` row into an OTLP span using the Sponsio Semantic
-    Conventions (see ``docs/reference/observability.md``), and writes them
-    either to a local OTLP-JSONL file or POSTs them to an OTLP/HTTP
-    collector (Datadog, Honeycomb, Grafana Cloud, the Sponsio-native
-    dashboard, …).
-
-    \b
-    Examples:
-      # Last 24h of audit, all agents, push to your dashboard
-      sponsio export-sessions --to https://obs.example.com/v1/traces \\
-                              --header "x-api-key: $OBS_API_KEY"
-
-      # Last 7d of one agent, write to a file
-      sponsio export-sessions --since 7d --agent _host_cursor \\
-                              --to ./audit-export.jsonl
-
-      # Everything we have, no time cutoff
-      sponsio export-sessions --since all --to ./full-audit.jsonl
-
-    The session log is the audit substrate (``MonitorEvent``-flat
-    records); the runtime span tree (per-phase precondition /
-    guarantee / sto_eval children) is *not* persisted to disk, so
-    historical exports are necessarily lossy on per-phase detail.
-    Live exports via :class:`sponsio.tracer.exporters.OtlpHttpExporter`
-    carry the full tree.
-    """
-    from sponsio.runtime.session_log import _resolve_default_base_dir
-
-    cutoff = _parse_since(since)
-    base = (
-        sessions_dir.expanduser()
-        if sessions_dir is not None
-        else _resolve_default_base_dir()
-    )
-
-    if not base.exists():
-        click.echo(
-            click.style(f"sessions dir not found: {base}", fg="yellow"),
-            err=True,
-        )
-        sys.exit(0)
-
-    # Walk per-agent subdirectories.
-    agent_dirs: list[Path]
-    if agent_filter is not None:
-        agent_dirs = [base / agent_filter]
-        if not agent_dirs[0].is_dir():
-            click.echo(
-                click.style(f"no sessions for agent {agent_filter!r}", fg="yellow"),
-                err=True,
-            )
-            sys.exit(0)
-    else:
-        agent_dirs = [p for p in base.iterdir() if p.is_dir()]
-
-    spans: list[dict] = []
-    by_agent: dict[str, int] = {}
-
-    for agent_dir in sorted(agent_dirs):
-        agent_id = agent_dir.name
-        for jsonl_path in sorted(agent_dir.glob("*.jsonl")):
-            try:
-                lines = jsonl_path.read_text().splitlines()
-            except OSError as e:
-                click.echo(
-                    click.style(f"  skip {jsonl_path}: {e}", fg="yellow"), err=True
-                )
-                continue
-            for ln in lines:
-                ln = ln.strip()
-                if not ln:
-                    continue
-                try:
-                    rec = json.loads(ln)
-                except json.JSONDecodeError:
-                    continue
-                if cutoff and float(rec.get("ts") or 0.0) < cutoff:
-                    continue
-                spans.append(_session_event_to_otlp_span(rec))
-                by_agent[agent_id] = by_agent.get(agent_id, 0) + 1
-
-    if not spans:
-        click.echo(
-            click.style(
-                f"no events matched (since={since}, agent={agent_filter})",
-                fg="yellow",
-            ),
-            err=True,
-        )
-        sys.exit(0)
-
-    # Emit one OTLP envelope.
-    from sponsio.tracer import semconv as _semconv
-
-    envelope = {
-        "resourceSpans": [
-            {
-                "resource": {
-                    "attributes": [
-                        _attr_for_session(
-                            "service.name",
-                            service_name or "sponsio-sessions",
-                        ),
-                    ],
-                },
-                "scopeSpans": [
-                    {
-                        "scope": {
-                            "name": "sponsio",
-                            "version": _semconv.SCHEMA_VERSION,
-                        },
-                        "schemaUrl": _semconv.SCHEMA_URL,
-                        "spans": spans,
-                    }
-                ],
-            }
-        ],
-    }
-
-    if destination.startswith(("http://", "https://")):
-        # HTTP push via the in-tree batching exporter.
-        headers: dict[str, str] = {}
-        for raw in headers_raw:
-            if ":" not in raw:
-                raise click.BadParameter(f"--header must be 'Key: Value' (got {raw!r})")
-            k, _, v = raw.partition(":")
-            headers[k.strip()] = v.strip()
-
-        body = json.dumps(envelope).encode("utf-8")
-        click.echo(
-            f"POSTing {len(spans)} spans ({len(body) / 1024:.1f} KB) → {destination}"
-        )
-        try:
-            req = urllib.request.Request(
-                destination,
-                data=body,
-                headers={"Content-Type": "application/json", **headers},
-                method="POST",
-            )
-            with urllib.request.urlopen(req, timeout=10.0) as resp:
-                if not (200 <= resp.status < 300):
-                    click.echo(
-                        click.style(
-                            f"collector returned HTTP {resp.status}",
-                            fg="red",
-                        ),
-                        err=True,
-                    )
-                    sys.exit(1)
-        except urllib.error.URLError as e:
-            click.echo(click.style(f"HTTP push failed: {e}", fg="red"), err=True)
-            sys.exit(1)
-        click.secho(f"✓ pushed {len(spans)} spans", fg="green")
-    else:
-        # File destination. write the OTLP envelope as a single JSON.
-        out = Path(destination).expanduser()
-        out.parent.mkdir(parents=True, exist_ok=True)
-        out.write_text(json.dumps(envelope, indent=2))
-        click.secho(
-            f"✓ wrote {len(spans)} spans → {out} ({out.stat().st_size / 1024:.1f} KB)",
-            fg="green",
-        )
-
-    # Summary by agent. useful when --agent isn't set.
-    if by_agent:
-        click.echo()
-        click.echo(click.style("By agent:", bold=True))
-        for agent_id, n in sorted(by_agent.items(), key=lambda x: -x[1]):
-            click.echo(f"  {agent_id:30}  {n:6} events")
-
-    click.echo()
-    click.echo(
-        click.style("Schema: ", dim=True)
-        + f"{_semconv.SCHEMA_URL} (version {_semconv.SCHEMA_VERSION})"
-    )
-
-
-@cli.command(name="eval")
-@click.argument(
-    "trace_path",
-    type=click.Path(exists=True, file_okay=True, dir_okay=True, path_type=Path),
-)
-@click.argument("contracts", nargs=-1)
-@click.option(
-    "--config",
-    "-c",
-    "config_path",
-    type=click.Path(exists=True),
-    help="YAML config file (sponsio.yaml)",
-)
-@click.option("--agent", "-a", "agent_id", help="Agent ID (with --config)")
-@click.option("--json", "as_json", is_flag=True, help="Output as JSON")
-@click.option(
-    "--baseline",
-    "baseline_path",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-    default=None,
-    help=(
-        "Diff against a previous JSON report (produced by `--json`).  "
-        "Surfaces FPR/FNR deltas per contract and overall."
-    ),
-)
-@click.option(
-    "--max-fpr-delta",
-    type=float,
-    default=None,
-    help=(
-        "Fail (exit 1) if overall FPR rose by more than this many "
-        "percentage points vs --baseline.  E.g. `0.01` = 1pp.  "
-        "Use in CI to catch overblock regressions automatically."
-    ),
-)
-@click.option(
-    "--max-fnr-delta",
-    type=float,
-    default=None,
-    help=(
-        "Fail (exit 1) if overall FNR rose by more than this many "
-        "percentage points vs --baseline.  Use to catch regressions "
-        "where contracts started missing real incidents."
-    ),
-)
-@click.option(
-    "--write-baseline",
-    "write_baseline_path",
-    type=click.Path(dir_okay=False, path_type=Path),
-    default=None,
-    help=(
-        "After running, write the report JSON to this path.  Use to "
-        "snapshot a green run as the new baseline for the next PR."
-    ),
-)
-def eval_cmd(
-    trace_path: Path,
-    contracts,
-    config_path,
-    agent_id,
-    as_json,
-    baseline_path: Path | None,
-    max_fpr_delta: float | None,
-    max_fnr_delta: float | None,
-    write_baseline_path: Path | None,
-):
-    """Replay a labelled trace corpus and report FPR/FNR per contract.
-
-    Use this BEFORE flipping ``SPONSIO_MODE=enforce``. it answers
-    "if I turn enforcement on tomorrow, how often will my contracts
-    over-block legitimate traffic, and how often will they miss real
-    incidents?".
-
-    Label convention: filename prefix.\n
-    \b
-        safe_login.json     → expected to PASS every contract
-        unsafe_drop.json    → expected to be BLOCKED by ≥1 contract
-        anything_else.json  → counted but not used in FPR/FNR
-
-    Examples:\n
-        sponsio eval traces/ --config sponsio.yaml --agent bot\n
-        sponsio eval traces/ "tool `transfer` at most 1 times"\n
-        sponsio eval traces/ --config sponsio.yaml --json\n
-        sponsio eval traces/ -c sponsio.yaml \\\n
-            --baseline main-baseline.json --max-fpr-delta 0.01
-
-    Reasonable CI gates: ``--max-fpr-delta 0.01`` (1pp overblock
-    regression budget) and ``--max-fnr-delta 0.0`` (zero tolerance
-    for new misses).  Adjust to your appetite.
-    """
-    from sponsio.eval_runner import (
-        diff_reports,
-        discover_cases,
-        format_diff,
-        format_report,
-        run_eval,
-    )
-
-    if config_path and contracts:
-        click.echo(
-            click.style(
-                "Error: cannot use both --config and positional contracts", fg="red"
-            )
-        )
-        sys.exit(1)
-    if agent_id and not config_path:
-        click.echo(click.style("Error: --agent requires --config", fg="red"))
-        sys.exit(1)
-    if not config_path and not contracts:
-        click.echo("Usage: sponsio eval TRACE_PATH [CONTRACTS...] [--config FILE]")
-        sys.exit(1)
-
-    # Resolve contracts to a flat list of NL strings / structured entries
-    contract_list: list = []
-    if config_path:
-        from sponsio.config import load_config
-
-        cfg = load_config(config_path)
-        if not agent_id:
-            if len(cfg.agents) == 1:
-                agent_id = next(iter(cfg.agents))
-            else:
-                click.echo(
-                    click.style(
-                        f"Error: multiple agents in config "
-                        f"({list(cfg.agents.keys())}), use --agent",
-                        fg="red",
-                    )
-                )
-                sys.exit(1)
-        for ce in cfg.agents[agent_id].contracts:
-            for field_value in (ce.assumption, ce.guarantee):
-                if field_value is None:
-                    continue
-                if isinstance(field_value, list):
-                    contract_list.extend(field_value)
-                else:
-                    contract_list.append(field_value)
-    else:
-        contract_list = list(contracts)
-
-    cases = discover_cases(trace_path)
-    if not cases:
-        click.echo(click.style(f"No trace files found at {trace_path}", fg="yellow"))
-        sys.exit(0)
-
-    report = run_eval(cases, contract_list)
-
-    # Validate flag combinations BEFORE doing the eval render so a
-    # typo doesn't cost the user a 30s replay.
-    if (max_fpr_delta is not None or max_fnr_delta is not None) and not baseline_path:
-        click.echo(
-            click.style(
-                "Error: --max-fpr-delta / --max-fnr-delta require --baseline",
-                fg="red",
-            )
-        )
-        sys.exit(2)
-
-    diff = None
-    if baseline_path:
-        try:
-            baseline_data = json.loads(baseline_path.read_text())
-        except (json.JSONDecodeError, OSError) as e:
-            click.echo(
-                click.style(f"Error reading baseline {baseline_path}: {e}", fg="red")
-            )
-            sys.exit(2)
-        diff = diff_reports(baseline_data, report)
-
-    if as_json:
-        # Preserve the long-standing flat shape (report fields at the
-        # top) when there's no baseline. every existing script
-        # depends on ``data["n_safe"]`` etc.  Only when a baseline
-        # IS present do we add a sibling key for the diff, which
-        # callers can look up only when they passed ``--baseline``.
-        out = report.to_dict()
-        if diff is not None:
-            out["baseline_diff"] = diff.to_dict()
-        click.echo(json.dumps(out, indent=2))
-    else:
-        click.echo(format_report(report))
-        if diff is not None:
-            click.echo(format_diff(diff))
-
-    # Snapshot the report for the next PR's --baseline.  Done AFTER
-    # the gate check so a regression-failing run doesn't auto-poison
-    # main's baseline (gate failures should not silently rewrite the
-    # standard you're being measured against).
-    gate_failures: list[str] = []
-    if diff is not None:
-        gate_failures = diff.gate_violations(
-            max_fpr_delta=max_fpr_delta,
-            max_fnr_delta=max_fnr_delta,
-        )
-        if gate_failures:
-            click.echo()
-            for v in gate_failures:
-                click.secho(f"  ✗ {v}", fg="red", bold=True)
-
-    if write_baseline_path and not gate_failures:
-        write_baseline_path.write_text(json.dumps(report.to_dict(), indent=2))
-        click.secho(f"\n  ✓ baseline written to {write_baseline_path}", fg="green")
-    elif write_baseline_path and gate_failures:
-        click.secho(
-            f"\n  · skipped writing {write_baseline_path} "
-            "(gate failed. fix the regression first)",
-            fg="yellow",
-        )
-
-    if gate_failures:
-        sys.exit(1)
-
-
-@cli.command()
-@click.argument(
-    "target",
-    type=click.Path(file_okay=True, dir_okay=True, path_type=Path),
-    default=".",
-    required=False,
-)
-@click.option(
-    "--plan",
-    "plan_spec",
-    default=None,
-    help=(
-        "Print the would-run commands for these picks, don't run them.  "
-        "Used by IDE-agent wizard prompts for the dry-run preview step."
-    ),
-)
-@click.option(
-    "--apply",
-    "apply_spec",
-    default=None,
-    help=(
-        "Run the commands for these picks non-interactively.  Picks "
-        "format: ``framework=<name>;ides=<ide>:<level>,<ide>:<level>;"
-        "mode=observe|enforce`` where ``<level>`` is one of ``none`` / "
-        "``skill`` / ``full``.  Legacy ``hosts=`` / ``skills=`` form is "
-        "still accepted (``hosts=X`` ↔ ``X:full``, ``skills=X`` ↔ "
-        "``X:skill``)."
-    ),
-)
-@click.option(
-    "--no-demo",
-    is_flag=True,
-    help="Skip the post-install demo offer.",
-)
-def init(
-    target: Path,
-    plan_spec: str | None,
-    apply_spec: str | None,
-    no_demo: bool,
-):
-    """Interactive 4-axis onboarding wizard.
-
-    Walks you through the four decisions that actually matter on
-    first run:
-
-    \b
-      1. framework wrap        which agent framework to wrap (or "none")
-      2. protect host agents   install hooks for claude-code / cursor / openclaw
-      3. install Sponsio skill drop SKILL.md into IDEs not picked above
-      4. mode                  observe (shadow, default) or enforce (block)
-
-    Three modes:
-
-    \b
-      sponsio init                     # interactive TTY (humans)
-      sponsio init --plan '<picks>'    # print commands, don't run
-      sponsio init --apply '<picks>'   # run commands non-interactively
-
-    Picks string format::
-
-        framework=<name>;ides=<ide>:<level>,<ide>:<level>;mode=<observe|enforce>
-
-    Where ``<level>`` is ``none`` / ``skill`` / ``full``.  Legacy
-    ``hosts=<a>,<b>;skills=<a>,<b>`` form is still accepted.
-
-    Examples:\n
-        sponsio init\n
-        sponsio init --plan 'framework=langgraph;ides=cursor:skill;mode=observe'\n
-        sponsio init --apply 'framework=langgraph;ides=cursor:skill;mode=observe'
-    """
-    from sponsio.init_wizard import (
-        apply_commands,
-        detect_environment,
-        offer_demo,
-        parse_picks,
-        plan_commands,
-        run_interactive,
-    )
-
-    if plan_spec is not None and apply_spec is not None:
-        raise click.UsageError("--plan and --apply are mutually exclusive")
-
-    target_dir = target if target.is_dir() or not target.suffix else target.parent
-    env = detect_environment(target_dir)
-    # Detect a pre-existing ``@sponsio/sdk`` install (via ``npm
-    # install`` OR ``npm link``) so we can skip the redundant
-    # install step in plan.  Skipping is critical for ``npm link``
-    # workflows. running ``npm install --save-dev`` against a
-    # linked package overwrites the symlink with the published
-    # release, silently undoing the user's local-source testing.
-    # The legacy ``@sponsio/scan-ts`` package is now a deprecation
-    # shim that re-exports ``@sponsio/sdk``'s CLI; counting it as
-    # "installed" lets users on the old name finish ``sponsio
-    # init`` without an extra install round-trip.
-    # The package dir alone isn't enough: a half-broken install can
-    # leave ``node_modules/@sponsio/sdk/`` populated but the bin
-    # symlink ``node_modules/.bin/sponsio`` missing. which would
-    # then make ``npx sponsio onboard`` fall through to the npm
-    # public registry (404, since pip-side ``sponsio`` shadows the
-    # name).  Require BOTH for the skip-install path.
-    _scan_ts_pkg = (target_dir / "node_modules" / "@sponsio" / "sdk").exists() or (
-        target_dir / "node_modules" / "@sponsio" / "scan-ts"
-    ).exists()
-    _scan_ts_bin = (target_dir / "node_modules" / ".bin" / "sponsio").exists()
-    _scan_ts_installed = _scan_ts_pkg and _scan_ts_bin
-
-    # ---- non-TTY paths: --plan / --apply ----
-    if plan_spec is not None:
-        picks = parse_picks(plan_spec)
-        cmds = plan_commands(
-            picks,
-            ts_project=env.runtime == "ts",
-            scan_ts_already_installed=_scan_ts_installed,
-        )
-        if not cmds:
-            click.echo(
-                "Nothing to do. picks select no framework wrap and no "
-                "IDE protection.  Re-run with at least one ``framework=`` "
-                "or ``ides=<ide>:full|skill``."
-            )
-            return
-        for cmd in cmds:
-            click.echo("would run: " + " ".join(cmd))
-        return
-
-    if apply_spec is not None:
-        picks = parse_picks(apply_spec)
-    else:
-        picks = run_interactive(env)
-
-    cmds = plan_commands(
-        picks,
-        ts_project=env.runtime == "ts",
-        scan_ts_already_installed=_scan_ts_installed,
-    )
-    if not cmds:
-        click.echo()
-        click.secho(
-            "Nothing to install. every axis was set to 'none'.  "
-            "Re-run `sponsio init` and pick at least one framework "
-            "wrap or IDE level (skill / full).",
-            fg="yellow",
-        )
-        return
-
-    # Dry-run preview before running, even on the interactive path.
-    # gives the user a final chance to spot a wrong pick.  Indented
-    # col-2 to match the wizard's body content margin (banner col-0,
-    # everything below at col-2).
-    click.echo()
-    click.secho("  preview", bold=True, fg="cyan")
-    for cmd in cmds:
-        click.echo("    → " + " ".join(cmd))
-    click.echo()
-
-    # Skip the confirm gate when called via ``--apply``. the IDE
-    # agent already showed me the dry-run preview before invoking,
-    # and a second confirmation here would corrupt structured output.
-    if apply_spec is None:
-        from sponsio.init_wizard import _confirm as _wizard_confirm
-
-        if not _wizard_confirm("Run these?", default=True):
-            click.echo()
-            click.secho("  ✘  No changes made.", fg="yellow")
-            click.echo(
-                "      Re-run `sponsio init` whenever you're ready, "
-                "or pass\n      `sponsio init --plan '<picks>'` to "
-                "preview the commands without prompts."
-            )
-            return
-
-    rc = apply_commands(cmds)
-    if rc != 0:
-        sys.exit(rc)
-
-    # Picks-aware "what now" block.  Each combination of axes leaves
-    # the user in a different spot. IDE-only installs especially
-    # were ending without any concrete next action.  Route through
-    # the helper so each path gets a tailored handoff.
-    from sponsio.init_wizard import print_next_steps as _print_next_steps
-
-    _print_next_steps(picks, ts_project=env.runtime == "ts")
-
-    if not no_demo:
-        offer_demo()
-
-
-# ---------------------------------------------------------------------------
-# refresh. re-mine contracts from recent traces and merge into sponsio.yaml
-# ---------------------------------------------------------------------------
-
-
-@cli.command()
-@click.argument(
-    "path",
-    type=click.Path(file_okay=True, dir_okay=True, path_type=Path),
-    default=".",
-    required=False,
-)
-@click.option(
-    "--llm",
-    is_flag=True,
-    help=(
-        "Make a real LLM call to verify connectivity, latency, and "
-        "credentials.  Opt-in because it costs a few tokens and ~1s; "
-        "default ``doctor`` is fully offline."
-    ),
-)
-@click.option(
-    "--json",
-    "as_json",
-    is_flag=True,
-    help=(
-        "Emit a structured JSON report instead of the human-readable "
-        "table.  Schema is stable per `schema_version`.  Use for "
-        "IDE integrations, CI gates, fleet dashboards, or piping into "
-        "`jq` / wrapper scripts."
-    ),
-)
-def doctor(path: Path, llm: bool, as_json: bool):
-    """Diagnose your Sponsio install and project wiring.
-
-    Runs a short battery of mostly-offline checks. Python version,
-    sponsio import sanity, optional SDK availability, LLM credentials,
-    ``sponsio.yaml`` validation, a project-level AST scan, and an
-    end-to-end guard smoke-test. and prints a single report telling
-    you exactly what to run next.
-
-    Pass ``--llm`` to also make a real LLM round-trip (uses the
-    provider/key from ``sponsio.yaml``'s ``extractor:`` section if
-    present, env-var auto-detection otherwise).
-
-    Exits non-zero if any check fails (warnings are advisory and don't
-    change the exit code), so ``doctor`` is safe to wire into CI as a
-    pre-flight sanity gate.
-
-    Examples:\n
-        sponsio doctor\n
-        sponsio doctor src/\n
-        sponsio doctor --llm\n
-        sponsio doctor path/to/sponsio.yaml --llm
-    """
-    from sponsio.doctor import print_report, report_to_dict, run_doctor
-
-    results, exit_code = run_doctor(path, with_llm=llm)
-    if as_json:
-        # Suppress the human-readable banner. JSON consumers want
-        # exactly one parseable document on stdout, nothing else.
-        click.echo(json.dumps(report_to_dict(results, exit_code), indent=2))
-    else:
-        print_report(results)
-    sys.exit(exit_code)
-
-
-# ---------------------------------------------------------------------------
-# bench
-# ---------------------------------------------------------------------------
-
-
-# ---------------------------------------------------------------------------
-# onboard
-# ---------------------------------------------------------------------------
-
-
-@cli.command()
-@click.argument(
-    "target",
-    type=click.Path(file_okay=True, dir_okay=True, path_type=Path),
-    default=".",
-    required=False,
-)
-@click.option(
-    "--agent",
-    "agent_id",
-    default="agent",
-    show_default=True,
-    help=(
-        "Agent identifier stamped into sponsio.yaml.  Matches "
-        "`sponsio scan`'s default so a later `scan --append` lands "
-        "in the same agent block."
-    ),
-)
-@click.option(
-    "--mode",
-    type=click.Choice(["observe", "enforce"]),
-    default=None,
-    help=(
-        "Runtime mode written into sponsio.yaml.  Skip the flag to be "
-        "prompted interactively (same Y/N question ``sponsio init`` "
-        "and ``sponsio host install`` ask).  ``observe`` is the safe "
-        "default. never blocks, logs every would-have-blocked decision "
-        "to ~/.sponsio/sessions/<agent_id>/*.jsonl."
-    ),
-)
-@click.option(
-    "--force",
-    is_flag=True,
-    help="Overwrite an existing sponsio.yaml without prompting.",
-)
-@click.option(
-    "--no-probe-ollama",
-    is_flag=True,
-    help=(
-        "Skip the localhost:11434 liveness probe.  Useful in CI or "
-        "behind strict firewalls where the <500ms probe still times "
-        "out slowly and you'd rather jump straight to the starter pack."
-    ),
-)
-@click.option(
-    "--no-doctor",
-    is_flag=True,
-    help=(
-        "Skip the post-onboard `sponsio doctor` run.  By default we "
-        "run the full offline check battery so users see whether the "
-        "install is healthy before they switch to enforce mode."
-    ),
-)
-@click.option(
-    "--json",
-    "as_json",
-    is_flag=True,
-    help="Emit the structured OnboardReport as JSON instead of text.",
-)
-@click.option(
-    "--emit-context",
-    "emit_context",
-    is_flag=True,
-    default=False,
-    help=(
-        "Skip the LLM step and instead emit the structured inputs "
-        "(framework / tool inventory / auto-selected packs / existing "
-        "yaml / discovered policy docs) as JSON to stdout. Used by the "
-        "host agent driving the ``sponsio`` skill: pair with "
-        "``sponsio prompt onboard`` and apply in the agent's own LLM "
-        "context. no UnifiedExtractor call, no extra API key."
-    ),
-)
-@click.option(
-    "--push/--no-push",
-    default=False,
-    help=(
-        "After writing sponsio.yaml, push it to the local dashboard at "
-        "--push-url so it lands on the Scan page + Contract Library "
-        "(default: off; on is one round-trip per run, silently skipped "
-        "when the dashboard isn't up)."
-    ),
-)
-@click.option(
-    "--push-url",
-    default="http://127.0.0.1:8000",
-    help="Dashboard URL to push to (default: http://127.0.0.1:8000).",
-)
-@click.option(
-    "--interactive/--no-interactive",
-    "interactive",
-    default=None,
-    help=(
-        "Prompt for framework / LLM provider / model up front and "
-        "write `.sponsiorc` + `.env.example` next to sponsio.yaml. "
-        "Default: auto. interactive when stdin is a TTY, "
-        "non-interactive otherwise (CI, scripts, docker entrypoints, "
-        "``--json``, ``--emit-context``).  Pass ``--no-interactive`` "
-        "to force the silent path even from a terminal."
-    ),
-)
-def onboard(
-    target: Path,
-    agent_id: str,
-    mode: str | None,
-    force: bool,
-    no_probe_ollama: bool,
-    no_doctor: bool,
-    as_json: bool,
-    emit_context: bool,
-    push: bool,
-    push_url: str,
-    interactive: bool | None,
-):
-    """One-shot project wire-up. detect framework, write sponsio.yaml, print patch.
-
-    Composes `init` + `scan` + `doctor` into a single command so
-    first-time users don't have to learn three subcommands just to
-    run the guard in observe mode.  Specifically:
-
-    \b
-      1. Detects the agent framework from imports + dependencies.
-      2. Detects the best available LLM provider (env vars →
-         OPENAI_BASE_URL → local Ollama → none).
-      3. Writes sponsio.yaml in observe mode with an inferred contract
-         set. LLM-inferred when a provider was found, or pure name-
-         heuristic starter pack when it wasn't.
-      4. Prints the framework-specific 2-line patch the user needs to
-         apply to their agent entry point.
-
-    Safe defaults throughout: mode=observe (never blocks on day 1),
-    agent_id="agent" (matches `sponsio scan`), and --force off (the
-    "I already have sponsio.yaml" case is louder than a silent overwrite).
-
-    Examples:\n
-        sponsio onboard\n
-        sponsio onboard src/\n
-        sponsio onboard . --agent customer_bot\n
-        sponsio onboard --force --no-probe-ollama
-    """
-    from sponsio.onboard import OnboardReport, run_onboard
-    from sponsio.runtime.spinner import Spinner
-
-    # Branded header. same ``header_banner`` Rich primitive that
-    # ``sponsio init`` / ``sponsio doctor`` / runtime print_banner /
-    # explain renderers use, so onboarding's first line of output
-    # looks like the rest of the CLI instead of a hand-glued
-    # ``━`` string.  Skipped on the non-interactive structured-
-    # output paths (--json, --emit-context) so consumers parsing
-    # stdout don't have to sed past it.
-    if not as_json and not emit_context and not os.environ.get("SPONSIO_INIT_DISPATCH"):
-        # Standalone ``sponsio onboard`` prints a full banner so users
-        # see the product wordmark when running it directly.  When the
-        # ``sponsio init`` wizard dispatched us, the preview block
-        # above already showed ``→ sponsio onboard . --mode ...`` so
-        # any additional banner / divider here is redundant.  Skip
-        # entirely and let the first stage section rule
-        # (``Scanning your code ─────...``) carry the transition.
-        from sponsio.render.components import header_banner as _header_banner
-        from sponsio.runtime.terminal import (
-            _make_stderr_console as _make_console,
-        )
-
-        _hdr_console = _make_console(None)
-        _hdr_console.print()
-        _hdr_console.print(_header_banner(tagline="onboard"))
-
-    # One spinner per command. long-wait emits (``…``-suffixed) start
-    # it, the next emit (or the final ``stop()`` after run_onboard)
-    # cleans up.  Skipped silently when stderr isn't a TTY, so CI / pipe
-    # / docker output stays line-oriented.
-    _spinner = Spinner()
-
-    def _progress(msg: str) -> None:
-        # ``▸`` prefix = stage section header.  Render as a
-        # :func:`section_rule` (label + ``─────...`` rule) so the
-        # divider matches the runtime trace renderer + ``sponsio init``
-        # axis headers.  Anything else is a per-step progress line.
-        # dim cyan ``· `` bullet.  Emits ending with ``…`` are "this
-        # will take a while" announcements; we hand them to the
-        # spinner so the user sees motion during the wait.
-        if as_json or emit_context:
-            return
-        # Always stop any running spinner first so the next line lands
-        # cleanly (rather than on top of a stale frame).
-        _spinner.stop()
-        if msg.startswith("▸ "):
-            from sponsio.render.components import (
-                indent as _indent_for_progress,
-                section_rule as _section_rule_for_progress,
-            )
-            from sponsio.runtime.terminal import (
-                _make_stderr_console as _make_console_for_progress,
-            )
-
-            _progress_console = _make_console_for_progress(None)
-            _progress_console.print()
-            _progress_console.print(
-                _indent_for_progress(_section_rule_for_progress(msg.removeprefix("▸ ")))
-            )
-            return
-        # Bullets indent col-2 to match the section rule above them
-        # and the trace + report renderers' body indentation.
-        line = "  " + click.style("· ", fg="cyan", dim=True) + msg
-        if msg.endswith("…"):
-            _spinner.start(line)
-        else:
-            click.echo(line, err=True)
-
-    # ---- agent-driven path: dump inputs, skip LLM step ------------------
-    # ``--emit-context`` runs the deterministic stages (framework /
-    # provider / AST tool inventory / pack selection) and stops short of
-    # the LLM contract-mining inside CodeAnalyzer.generate_yaml.  The
-    # host agent picks up where we leave off using ``sponsio prompt
-    # onboard``.
-    if emit_context:
-        target_path = Path(target)
-        if target_path.suffix in {".yaml", ".yml"}:
-            root = target_path.parent or Path(".")
-            existing_yaml_path = target_path
-        else:
-            root = target_path
-            existing_yaml_path = target_path / "sponsio.yaml"
-
-        from sponsio.discovery.extractors.code_analysis import CodeAnalyzer
-        from sponsio.onboard import detect_framework, select_packs
-
-        # AST-only. explicit ``use_llm=False`` so this path never
-        # reads any provider env var.
-        analyzer = CodeAnalyzer(use_llm=False)
-        tool_inventory = analyzer.get_tool_inventory([str(root)]) or []
-        # Run framework detection AFTER tool inventory, prioritizing
-        # the files the extractor already pinned as agent code. fixes
-        # the monorepo case where 200+ pad files at the root could
-        # exhaust the framework scan cap before the agent file (in a
-        # deep subdir) was reached, leaving framework="none" even
-        # though tool_inventory had found ``@tool`` functions there.
-        prioritize_files: list[Path] = []
-        for t in tool_inventory:
-            fp = t.get("filepath") if isinstance(t, dict) else None
-            if not fp:
-                continue
-            p = Path(fp)
-            if not p.is_absolute():
-                p = root / p
-            if p.is_file():
-                prioritize_files.append(p)
-        framework = detect_framework(root, prioritize_files=prioritize_files)
-        pack_selection = select_packs(framework.framework, tool_inventory)
-
-        existing_yaml_text = ""
-        if existing_yaml_path.exists():
-            try:
-                existing_yaml_text = existing_yaml_path.read_text(encoding="utf-8")
-            except OSError:
-                pass
-
-        # Surface common policy docs the agent should weight in the
-        # extraction.  Conservative search. root-level only, by
-        # convention. to avoid pulling in unrelated repo prose.  Dedup
-        # by inode so case-insensitive filesystems (macOS HFS+) don't
-        # report ``security.md`` and ``SECURITY.md`` twice.
-        policy_docs = []
-        seen_inodes: set[tuple[int, int]] = set()
-        for candidate in ("security.md", "SECURITY.md", "policy.md", "POLICY.md"):
-            p = root / candidate
-            if not p.is_file():
-                continue
-            try:
-                stat = p.stat()
-                key = (stat.st_dev, stat.st_ino)
-                if key in seen_inodes:
-                    continue
-                seen_inodes.add(key)
-                policy_docs.append(
-                    {
-                        "path": str(p.relative_to(root)),
-                        "content": p.read_text(encoding="utf-8"),
-                    }
-                )
-            except OSError:
-                pass
-
-        # Pull the framework-specific wrap snippet (the 2-3 line patch
-        # the user pastes into their agent entry file).  The skill's
-        # W1 step 5 references this field; emitting it here lets the
-        # agent surface the wiring instructions in the same turn it
-        # writes the YAML.
-        wrap_snippet_text = ""
-        try:
-            from sponsio.onboard import _wrap_snippet  # type: ignore[attr-defined]
-
-            wrap_snippet_text = _wrap_snippet(framework.framework, agent_id) or ""
-        except Exception:  # pragma: no cover. best-effort
-            pass
-
-        # Locate likely agent entry files so the IDE agent doesn't have
-        # to re-discover them. Conservative regex grep over root-level
-        # .py files, ranked by signal density.
-        entry_file_candidates: list[dict] = []
-        try:
-            framework_signals: dict[str, list[re.Pattern]] = {
-                "langchain": [
-                    re.compile(r"from\s+langchain"),
-                    re.compile(r"create_react_agent\s*\("),
-                ],
-                "langgraph": [
-                    re.compile(r"from\s+langgraph"),
-                    re.compile(r"StateGraph\s*\("),
-                    re.compile(r"create_react_agent\s*\("),
-                ],
-                "crewai": [re.compile(r"from\s+crewai"), re.compile(r"\bAgent\s*\(")],
-                "autogen": [
-                    re.compile(r"from\s+autogen"),
-                    re.compile(r"AssistantAgent\s*\("),
-                ],
-                "openai_agents": [
-                    re.compile(r"from\s+agents"),
-                    re.compile(r"\bAgent\s*\("),
-                ],
-                "openai": [re.compile(r"from\s+openai"), re.compile(r"OpenAI\s*\(")],
-                "anthropic": [
-                    re.compile(r"from\s+anthropic"),
-                    re.compile(r"Anthropic\s*\("),
-                    re.compile(r"messages\.create\s*\("),
-                ],
-                "claude_agent_sdk": [re.compile(r"from\s+claude_agent_sdk")],
-                "google_adk": [re.compile(r"from\s+google\.adk")],
-            }
-            sigs = framework_signals.get(framework.framework, [])
-            if sigs:
-                from glob import glob as _glob
-
-                py_files = sorted(
-                    set(
-                        _glob(str(root / "*.py"))
-                        + _glob(str(root / "**/*.py"), recursive=True)
-                    )
-                )
-                py_files = [
-                    f
-                    for f in py_files
-                    if "/.venv/" not in f
-                    and "/__pycache__/" not in f
-                    and "/site-packages/" not in f
-                ]
-                scored = []
-                for f in py_files[:200]:  # cap to avoid scanning large monorepos
-                    try:
-                        text = Path(f).read_text(encoding="utf-8")
-                    except OSError:
-                        continue
-                    matches = [s.pattern for s in sigs if s.search(text)]
-                    if matches:
-                        scored.append(
-                            {
-                                "path": str(Path(f).relative_to(root)),
-                                "reason": "matches: " + ", ".join(matches),
-                            }
-                        )
-                scored.sort(key=lambda x: -len(x["reason"]))
-                entry_file_candidates = scored[:5]
-        except Exception:  # pragma: no cover. best-effort
-            entry_file_candidates = []
-
-        # Parse the on-disk sponsio.yaml's contracts (if any) so the
-        # host agent driving this can dedupe its semantic-pass
-        # proposals without having to re-grep YAML.  Conservative
-        # parse. failures degrade to an empty list rather than
-        # blocking the emit (a malformed yaml is worth surfacing,
-        # but not at the cost of also blocking the rest of the
-        # diagnostic JSON).
-        pre_existing_contracts: list[dict] = []
-        if existing_yaml_path.exists():
-            pre_existing_contracts = _parse_existing_contracts(
-                existing_yaml_path, agent_id
-            )
-
-        # Health flag. the host agent uses this as the single
-        # gate for "should I keep going or stop and ask?".
-        # Reflects three orthogonal failure modes the previous
-        # case-A/B/C check in the wizard prompt was hand-rolling.
-        if framework.framework != "none" and tool_inventory:
-            health = "ok"
-            health_detail = "framework + tools detected"
-        elif tool_inventory:
-            # Rare after the prioritize-files fix, but still possible
-            # for unusual import shapes (star imports, dynamic
-            # `__import__`, etc.). surface explicitly so the agent
-            # asks the user to pick from axis 1 manually.
-            health = "tools_only"
-            health_detail = (
-                "tool_inventory found tools but no framework import "
-                "matched any known signature. pick framework manually"
-            )
-        elif framework.framework != "none":
-            health = "tools_only"
-            health_detail = (
-                f"framework {framework.framework!r} detected but "
-                "tool_inventory is empty. agent likely uses external "
-                "SDK tools (MCP, prebuilt LangChain tools, OpenAI "
-                "JSON schemas); grep the repo for tool registration"
-            )
-        else:
-            health = "empty"
-            health_detail = (
-                "no framework, no tools. wrong scan path "
-                "(monorepo + agent in subdir), or this is a bare "
-                "function-calling loop, or the project is TS and "
-                "you ran the Python probe"
-            )
-
-        click.echo(
-            json.dumps(
-                {
-                    "health": health,
-                    "health_detail": health_detail,
-                    "framework": {
-                        "name": framework.framework,
-                        "evidence": framework.evidence,
-                    },
-                    "agent_id": agent_id,
-                    "tool_inventory": tool_inventory,
-                    "auto_selected_packs": list(pack_selection.packs),
-                    "needs_workspace": pack_selection.needs_workspace,
-                    "existing_yaml": existing_yaml_text,
-                    "pre_existing_contracts": pre_existing_contracts,
-                    "policy_docs": policy_docs,
-                    "wrap_snippet": wrap_snippet_text,
-                    "entry_file_candidates": entry_file_candidates,
-                    "out_path": str(existing_yaml_path),
-                    "next_steps_hint": (
-                        "Run ``sponsio prompt onboard`` to get the prompt "
-                        "template, apply it to this JSON in your own LLM "
-                        "context, then write the resulting YAML to "
-                        f"{existing_yaml_path} via Edit/Write, and patch "
-                        "the agent entry file (see entry_file_candidates) "
-                        "with the wrap_snippet."
-                    ),
-                },
-                indent=2,
-                ensure_ascii=False,
-            )
-        )
-        return
-
-    # ---- interactive setup (prompts + dotfile writes) ------------------
-    # Decide whether to run prompts.  --json and --emit-context force
-    # non-interactive (prompts would corrupt the structured output).
-    # Otherwise an explicit --interactive / --no-interactive flag wins;
-    # without one, follow the TTY: real shell → prompts, CI / pipe /
-    # docker entrypoint → silent.
-    from sponsio.onboard import _wrap_snippet  # type: ignore[attr-defined]
-    from sponsio.onboard import detect_framework as _detect_fw_for_prompts
-    from sponsio.onboard import detect_provider as _detect_prov_for_prompts
-    from sponsio.onboard_setup import (
-        SetupAnswers,
-        maybe_no_api_key_warning,
-        run_setup_prompts,
-        stdin_is_tty,
-        write_sponsiorc,
-    )
-    from sponsio.sponsiorc import load_sponsiorc
-
-    if as_json or emit_context:
-        is_interactive = False
-    elif interactive is not None:
-        is_interactive = interactive
-    else:
-        is_interactive = stdin_is_tty()
-
-    # Resolve runtime mode through the same shared helper that
-    # ``sponsio host install`` uses, so all install paths ask the
-    # observe-vs-enforce question the same way. ``--mode`` skips the
-    # prompt; ``--json`` / ``--emit-context`` / ``--no-interactive``
-    # also skip it (structured-output paths must not pollute stdout
-    # with a click prompt). Fallback when no signal: ``observe``.
-    mode_was_explicit = mode is not None
-    mode = _resolve_runtime_mode(mode, allow_prompt=is_interactive)
-
-    target_dir = target if target.is_dir() else target.parent
-
-    # Resolve where sponsio.yaml will live so we can detect a "second
-    # run" case below without duplicating run_onboard's path logic.
-    if target.suffix in {".yaml", ".yml"}:
-        out_path_check = target
-    else:
-        out_path_check = target_dir / "sponsio.yaml"
-    yaml_already_exists = out_path_check.exists() and not force
-
-    # Second-run UX: if the user already ran onboard here (.sponsiorc is
-    # present), skip the prompts and reuse the saved choices.  Re-asking
-    # every time was annoying and the user explicitly flagged it.
-    rc = load_sponsiorc(target_dir) if target_dir.exists() else None
-    rc_in_target = (
-        rc is not None
-        and rc.found
-        and rc.source_path is not None
-        and rc.source_path.parent.resolve() == target_dir.resolve()
-    )
-
-    if rc_in_target:
-        # Reuse the rcfile values verbatim. that's the whole point of
-        # the dotfile.  Prompts only fire when there's nothing to reuse.
-        # We still run framework detection so the wrap snippet on the
-        # yaml-preserve path reflects current code (not a stale
-        # rcfile).  Detection beating rcfile here is intentional: the
-        # only way ``framework`` ends up wrong in an rcfile is when an
-        # older detection couldn't recognise the user's code; if today's
-        # detector finds something concrete, that's the better answer.
-        pre_fw = _detect_fw_for_prompts(target_dir) if target_dir.exists() else None
-        detected_fw = (
-            pre_fw.framework if pre_fw and pre_fw.framework != "none" else None
-        )
-        answers = SetupAnswers(
-            framework=detected_fw or rc.framework or "none",
-            provider=rc.extractor_provider or "none",
-            model=rc.extractor_model or "",
-            api_key_env=rc.extractor_api_key_env or "",
-        )
-        pre_prov = None
-    else:
-        # Pre-detect framework + provider so the prompts have sensible
-        # defaults.  Cheap (no LLM); run even in non-interactive mode
-        # so the rcfile we write below reflects what onboard actually
-        # used.
-        pre_fw = _detect_fw_for_prompts(target_dir) if target_dir.exists() else None
-        pre_prov = _detect_prov_for_prompts(probe_ollama=not no_probe_ollama)
-        answers = run_setup_prompts(
-            detected_framework=pre_fw.framework if pre_fw else "none",
-            detected_provider=pre_prov.provider,
-            detected_model=pre_prov.model or "",
-            detected_api_key_env=pre_prov.env_var or "",
-            interactive=is_interactive,
-        )
-
-    # Second-run UX: existing sponsio.yaml + no --force → preserve it.
-    # We still refresh the dotfiles + reprint the wrap snippet so the
-    # command stays useful (re-running onboard to remind yourself how
-    # to wire it up shouldn't error).  --force keeps the regenerate
-    # path for users who actually want a fresh yaml.
-    report: OnboardReport | None = None
-    if yaml_already_exists:
-        if not as_json and not emit_context:
-            click.echo()
-            click.secho(f"✓ {out_path_check}", fg="green")
-            click.echo("  preserved (re-run with --force to regenerate)")
-    else:
-        try:
-            report = run_onboard(
-                target,
-                agent_id=agent_id,
-                mode=mode,
-                force=force,
-                probe_ollama=not no_probe_ollama,
-                run_doctor=not no_doctor,
-                progress=_progress,
-            )
-        except FileExistsError as e:
-            _spinner.stop()
-            click.echo(click.style("Error: ", fg="red") + str(e), err=True)
-            sys.exit(1)
-        finally:
-            # Belt + braces: if the last emit was a ``…`` line (rare.
-            # run_onboard normally pairs each "Running …" with a "done"
-            # emit), make sure we don't leave the spinner thread spinning
-            # forever and the cursor stuck on a stale frame.
-            _spinner.stop()
-
-    # Write the rcfile (idempotent, plain write_text).  Skipped when
-    # target was a single file rather than a directory. the rcfile
-    # location is ambiguous in that case.  We deliberately do NOT
-    # write a ``.env.example`` here: sponsio reads ``os.environ``
-    # directly (no python-dotenv in the runtime), so a ``.env``-based
-    # recipe would silently fail.  Users keep secrets in their shell
-    # rc / direnv / system keychain. the rcfile records only the
-    # variable name (``api_key_env``), not the value.
-    sponsiorc_path: Path | None = None
-    if target_dir.exists() and target_dir.is_dir():
-        sponsiorc_path = write_sponsiorc(answers, target_dir)
-
-    if as_json:
-        payload = (
-            report.to_dict()
-            if report is not None
-            else {
-                "out_path": str(out_path_check),
-                "preserved": True,
-            }
-        )
-        payload["setup"] = {
-            "interactive": is_interactive,
-            "framework": answers.framework,
-            "provider": answers.provider,
-            "model": answers.model,
-            "api_key_env": answers.api_key_env,
-            "api_key_set_in_env": answers.api_key_set_in_env,
-            "sponsiorc_path": str(sponsiorc_path) if sponsiorc_path else None,
-        }
-        click.echo(json.dumps(payload, indent=2))
-        return
-
-    # Human-readable summary.  Kept compact so the wrap snippet is the
-    # last thing the user sees. it's what they need to act on.  When
-    # report is None we're on the second-run preserve path; the "✓
-    # sponsio.yaml preserved" line was already printed above.
-    if report is not None:
-        click.echo()
-        click.secho(f"  ✓ {report.out_path}", fg="green")
-        click.echo(f"      tools:      {report.tools_count}")
-        click.echo(f"      contracts:  {report.contracts_count}")
-        click.echo(f"      mode:       {report.mode}")
-        click.echo(f"      framework:  {report.framework.framework}")
-        click.echo(f"      provider:   {report.provider.provider}")
-        if report.starter_pack_used:
-            click.secho(
-                "      · starter-pack applied (no-LLM safety net)",
-                fg="yellow",
-                dim=True,
-            )
-
-    # Dotfiles written alongside sponsio.yaml.  Surface the paths so
-    # the user knows which file holds their tool config (vs. the
-    # contract library) and where to drop their actual API key.
-    if sponsiorc_path is not None:
-        click.echo()
-        click.secho(f"  ✓ {sponsiorc_path}", fg="green")
-        click.echo(
-            "      framework + LLM config. edit this file to change "
-            "framework / model / api_key_env"
-        )
-        # Best-effort .gitignore hint: only fire when sponsiorc is in
-        # a git repo AND `.sponsiorc` isn't already covered by the
-        # existing rules.  Avoids nagging users who already gitignore'd
-        # it (or who deliberately track it for team-wide config).
-        try:
-            rc_dir = sponsiorc_path.parent
-            git_root = rc_dir
-            for _ in range(8):  # walk up to 8 levels. plenty for a repo
-                if (git_root / ".git").exists():
-                    break
-                if git_root.parent == git_root:
-                    git_root = None  # type: ignore[assignment]
-                    break
-                git_root = git_root.parent
-            else:
-                git_root = None
-            if git_root is not None:
-                gitignore = git_root / ".gitignore"
-                already_ignored = False
-                if gitignore.is_file():
-                    ignore_text = gitignore.read_text(encoding="utf-8")
-                    for line in ignore_text.splitlines():
-                        s = line.strip()
-                        if s and not s.startswith("#"):
-                            if s in {".sponsiorc", "**/.sponsiorc", "*.sponsiorc"}:
-                                already_ignored = True
-                                break
-                if not already_ignored:
-                    click.secho(
-                        "  tip: add `.sponsiorc` to .gitignore "
-                        "(holds local model / api_key_env hints)",
-                        fg="cyan",
-                        dim=True,
-                    )
-        except OSError:
-            pass
-    # No-key warning. fires when the user picked a provider that
-    # needs a key but the env var isn't actually set, or when
-    # provider==none (so onboard fell back to the name-heuristic
-    # starter pack instead of LLM-inferred contracts).
-    no_key_msg = maybe_no_api_key_warning(answers)
-    if no_key_msg is not None:
-        click.echo()
-        for ln in no_key_msg.splitlines():
-            click.secho("  " + ln, fg="yellow")
-
-    if report is not None and report.doctor_results is not None:
-        total = len(report.doctor_results)
-        fails = sum(1 for r in report.doctor_results if r.status == "fail")
-        warns = sum(1 for r in report.doctor_results if r.status == "warn")
-        if fails == 0 and warns == 0:
-            click.secho(f"  ✓ doctor:   {total}/{total} checks passed", fg="green")
-        else:
-            click.echo(
-                f"  doctor:     {total - fails - warns}/{total} ok"
-                + (click.style(f", {warns} warn", fg="yellow") if warns else "")
-                + (click.style(f", {fails} fail", fg="red") if fails else "")
-            )
-            for r in report.doctor_results:
-                if r.status in {"fail", "warn"}:
-                    color = "red" if r.status == "fail" else "yellow"
-                    click.echo(
-                        f"    {click.style(r.icon, fg=color)} {r.name}: {r.detail}"
-                    )
-
-    if report is not None:
-        for w in report.warnings:
-            click.echo()
-            click.echo(click.style("  warn: ", fg="yellow") + w)
-
-    # Print the framework-specific patch snippet.  Auto-applying it
-    # to the user's agent file used to live behind ``--apply`` but
-    # was removed (only langgraph / langchain were supported, and a
-    # coding agent / manual paste does the same job for any
-    # framework with fewer surprises).  On the second-run preserve
-    # path the framework comes from the rcfile-derived answers (the
-    # user's saved choice, not a fresh detection).
-    snippet = (
-        report.wrap_snippet
-        if report is not None
-        else _wrap_snippet(answers.framework or "none", agent_id)
-    )
-    click.echo()
-    click.secho("Add this to your agent entry point:", bold=True)
-    click.echo()
-    for ln in snippet.splitlines():
-        click.echo(f"  {click.style(ln, fg='cyan')}")
-
-    # Surface the contract file the user should now review. ``onboard``
-    # wrote LLM-inferred (or starter-pack) contracts based on detected
-    # tools. they're a sane first cut, not a finished policy. Pointing
-    # the user at the path with a clear "review before flipping to
-    # enforce" callout turns "did onboard actually do what I wanted?"
-    # into one ``cat`` command.
-    review_path = (
-        report.out_path
-        if report is not None
-        else (out_path_check if yaml_already_exists else None)
-    )
-    if review_path is not None and not as_json and not emit_context:
-        click.echo()
-        click.secho("Review the generated contracts:", bold=True)
-        click.echo(f"  {click.style(str(review_path), fg='green')}")
-        click.secho(
-            "  (open it, sanity-check each rule, then re-run with `--mode enforce`",
-            dim=True,
-        )
-        click.secho("   when you're ready to switch from observe to active)", dim=True)
-
-    # --push: surface the generated yaml in the local dashboard (one
-    # command == everything the dashboard needs). Silently skipped if
-    # the dashboard isn't running, so a CI invocation without `serve`
-    # up doesn't fail.
-    if push and report is not None:
-        try:
-            yaml_content = report.out_path.read_text()
-        except Exception as e:
-            click.echo(
-                click.style("\n  push skipped: ", fg="yellow")
-                + f"could not read {report.out_path} ({e})"
-            )
-        else:
-            click.echo()
-            _push_scan_to_dashboard(
-                yaml_content=yaml_content,
-                filename=report.out_path.name,
-                dashboard_url=push_url,
-                source_paths=[str(target)],
-            )
-
-    # Optional immediate flip-to-enforce prompt.  Onboard always
-    # writes ``mode: observe`` by default. that's the safe path for
-    # teams who want a soak period.  But some users (CI hardening
-    # workflows, demo recordings, "I already ran the agent and know
-    # the contracts are right") want enforce on day 1.  Asking here
-    # turns "remember to sed the yaml later" into one keystroke.
-    #
-    # Skipped when:
-    #   - non-interactive (no TTY / --no-interactive / --json /
-    #     --emit-context. prompts would corrupt structured output)
-    #   - the user already chose ``--mode enforce`` (no point asking
-    #     a question they answered on the command line)
-    #   - run_onboard didn't actually produce a report (early-exit
-    #     paths above)
-    if (
-        report is not None
-        and is_interactive
-        and not as_json
-        and not emit_context
-        and not mode_was_explicit  # honor caller's `--mode` choice
-        and not os.environ.get("SPONSIO_INIT_DISPATCH")
-        and report.mode == "observe"
-    ):
-        click.echo()
-        flip = click.confirm(
-            click.style(
-                "Mode is `observe` (shadow). Flip to `enforce` now?",
-                bold=True,
-            ),
-            default=False,
-            show_default=True,
-        )
-        if flip:
-            try:
-                yaml_text = report.out_path.read_text(encoding="utf-8")
-                # Defer to the shared helper so both the interactive
-                # onboard flow and the explicit ``sponsio mode`` CLI
-                # agree on which mode-line to patch (runtime preferred,
-                # defaults fallback, append-observe-only as last resort).
-                new_yaml, action = _patch_mode_in_yaml(yaml_text, "enforce")
-                if action == "unchanged":
-                    click.secho(
-                        f"  ✓ {report.out_path} is already `mode: enforce`",
-                        fg="green",
-                    )
-                elif action == "missing":
-                    # ``onboard`` always writes a defaults.mode line so
-                    # this branch is defensive; surface it as a
-                    # yellow warning rather than overwriting.
-                    click.secho(
-                        f"  ✗ no `mode:` line in {report.out_path}, leaving "
-                        f"as observe. Add a `runtime: mode: enforce` block "
-                        f"by hand to flip.",
-                        fg="yellow",
-                    )
-                else:
-                    report.out_path.write_text(new_yaml, encoding="utf-8")
-                    suffix = (
-                        " (appended runtime: block)"
-                        if action == "appended"
-                        else f" ({action}.mode)"
-                    )
-                    click.secho(
-                        f"  ✓ flipped {report.out_path} → mode: enforce{suffix}",
-                        fg="green",
-                    )
-            except OSError as e:
-                click.secho(f"  ✗ could not rewrite {report.out_path}: {e}", fg="red")
-
-    click.echo()
-    click.echo("Next:")
-    click.echo("  sponsio report --since 24h            # what would have been blocked")
-    click.echo(
-        "  sponsio mode enforce                  # one-shot flip when you're ready"
-    )
-    click.echo()
-
-
-# ---------------------------------------------------------------------------
-# `sponsio plugin ...`. host-plugin runtime adapter
-# ---------------------------------------------------------------------------
-# `sponsio prompt <flow>`. workflow prompts for host-agent driving
-# ---------------------------------------------------------------------------
-#
-# Counterpart to ``sponsio plugin prompt <host>``: prints the agent-facing
-# extraction prompt for a top-level workflow (``onboard`` / ``scan``).
-# The setup skill at ``sponsio/skills/sponsio/SKILL.md`` calls this so the
-# host agent (Claude Code, Cursor, Codex) can apply the prompt in its own
-# LLM context against the JSON emitted by ``sponsio onboard --emit-context``
-# or ``sponsio scan --emit-context``.
-
-
-def _patch_mode_in_yaml(text: str, target_mode: str) -> tuple[str, str]:
-    r"""Set ``mode:`` to ``target_mode`` in a ``sponsio.yaml`` text body.
-
-    The Python loader reads ``runtime.mode`` AND ``defaults.mode``;
-    the TS loader reads only ``runtime.mode``. To stay correct under
-    both, this prefers the ``runtime.mode`` line, then falls back to
-    ``defaults.mode``, then appends a fresh ``runtime:`` block when
-    neither exists.
-
-    A naive ``re.subn(r"^\s*mode:")`` is wrong: it patches whichever
-    ``mode:`` line happens to come first in the file. If a yaml has
-    both ``runtime.mode`` and ``defaults.mode`` and ``defaults`` is
-    listed first, the runtime line silently stays stale and the TS
-    loader keeps reading the old value. Walking parent keys avoids
-    that whole class of bug.
-
-    Safety policy on appending: when neither block exists, this helper
-    will append a fresh ``runtime:`` block ONLY for ``target_mode ==
-    "observe"`` (the safe default). For ``target_mode == "enforce"``
-    against a malformed / missing-mode yaml, it returns ``"missing"``
-    instead of writing, so callers can refuse to silently flip a yaml
-    they cannot verify into the blocking posture. This matches the
-    OWASP-style principle of explicit-opt-in for enforcement, and
-    preserves CI scripts that relied on the old exit-1 behaviour for
-    malformed configs.
-
-    Returns:
-        ``(new_text, action)`` where ``action`` is one of
-        ``"runtime"`` / ``"defaults"`` / ``"appended"`` / ``"unchanged"``
-        / ``"missing"``. ``"unchanged"`` means the file already had
-        the desired value. ``"missing"`` means no mode line exists and
-        target is ``enforce`` so the helper refused to append.
-    """
-    lines = text.splitlines(keepends=True)
-    current_parent: str | None = None
-    runtime_idx = -1
-    defaults_idx = -1
-    mode_line_re = re.compile(r"^(\s+)mode:\s*(observe|enforce)(\s*(?:#.*)?)$")
-
-    for i, raw in enumerate(lines):
-        line = raw.rstrip("\n").rstrip("\r")
-        stripped = line.strip()
-        if not stripped or stripped.startswith("#"):
-            continue
-        indent = len(line) - len(line.lstrip())
-        if indent == 0:
-            # Top-level key like ``runtime:`` / ``defaults:`` / ``agents:``.
-            # Track only the most recent top-level key; only valid yaml
-            # mappings reach here so ``key:`` parsing is safe.
-            if ":" in stripped:
-                current_parent = stripped.split(":", 1)[0].strip()
-            continue
-        if mode_line_re.match(line):
-            if current_parent == "runtime" and runtime_idx < 0:
-                runtime_idx = i
-            elif current_parent == "defaults" and defaults_idx < 0:
-                defaults_idx = i
-
-    target_idx = runtime_idx if runtime_idx >= 0 else defaults_idx
-    if target_idx >= 0:
-        action = "runtime" if target_idx == runtime_idx else "defaults"
-        raw = lines[target_idx]
-        new_raw = re.sub(
-            r"^(\s+mode:\s*)(observe|enforce)(\s*(?:#.*)?)$",
-            lambda m: f"{m.group(1)}{target_mode}{m.group(3)}",
-            raw.rstrip("\n"),
-        )
-        # Preserve original line-ending.
-        ending = raw[len(raw.rstrip("\n")) :]
-        lines[target_idx] = new_raw + ending
-        new_text = "".join(lines)
-        if new_text == text:
-            return text, "unchanged"
-        return new_text, action
-
-    # Neither block had a mode line. Append only for the safe default
-    # (observe). Refuse to materialise an enforce block out of thin
-    # air: a missing mode line is suspicious, and silently flipping
-    # such a yaml into the blocking posture would mask CI / config
-    # errors that the operator should fix by hand.
-    if target_mode != "observe":
-        return text, "missing"
-    # Using ``runtime:`` (not ``defaults:``) so the TS loader picks it
-    # up too. Trailing newline normalised so the appended block doesn't
-    # glue to the previous line.
-    suffix = "" if text.endswith("\n") or text == "" else "\n"
-    appended = f"{suffix}\nruntime:\n  mode: {target_mode}\n"
-    return text + appended, "appended"
-
-
-@cli.command(name="mode")
-@click.argument(
-    "target_mode",
-    metavar="MODE",
-    type=click.Choice(["observe", "enforce"]),
-)
-@click.option(
-    "--config",
-    "-c",
-    "config_path",
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
-    default=Path("sponsio.yaml"),
-    show_default=True,
-    help="Path to the sponsio.yaml whose mode should be flipped.",
-)
-def cmd_mode(target_mode: str, config_path: Path):
-    """Flip a sponsio.yaml between `observe` and `enforce` in one shot.
-
-    The expected workflow is:
-
-    \b
-        sponsio onboard .            # writes sponsio.yaml in observe
-        # ...soak in observe for a day or two, watch `sponsio report`...
-        sponsio mode enforce         # one-line flip when you're ready
-
-    Prefers to update the ``runtime.mode:`` line (which both the Python
-    and TS loaders read), falling back to ``defaults.mode:`` (Python
-    only) or inserting a fresh ``runtime:`` block when neither exists.
-    Comments and surrounding lines survive untouched.
-    """
-    text = config_path.read_text(encoding="utf-8")
-    new_text, action = _patch_mode_in_yaml(text, target_mode)
-    if action == "unchanged":
-        click.echo(
-            click.style(
-                f"✓ {config_path} is already `mode: {target_mode}` (no change)",
-                fg="green",
-                dim=True,
-            )
-        )
-        return
-    if action == "missing":
-        # No mode line in the yaml AND target is enforce. Refuse to
-        # create one out of thin air. Tell the operator how to fix it
-        # safely (run observe first, edit yaml, or re-run onboard).
-        click.echo(
-            click.style(
-                f"✗ no `mode:` line found in {config_path} and refusing to "
-                f"append a fresh ``runtime: mode: enforce`` block. Run "
-                f"`sponsio mode observe` first to materialise the block, "
-                f"then `sponsio mode enforce` to flip it. Or edit the yaml "
-                f"by hand, or re-run `sponsio onboard --force`.",
-                fg="red",
-            ),
-            err=True,
-        )
-        raise SystemExit(1)
-    config_path.write_text(new_text, encoding="utf-8")
-    if action == "appended":
-        click.echo(
-            click.style("✓ ", fg="green")
-            + f"{config_path} → mode: {target_mode} (appended runtime: block)"
-        )
-    else:
-        click.echo(
-            click.style("✓ ", fg="green")
-            + f"{config_path} → {action}.mode: {target_mode}"
-        )
-
-
-@cli.command(name="prompt")
-@click.argument(
-    "flow",
-    type=click.Choice(["onboard", "scan"]),
-)
-def cmd_prompt(flow: str):
-    """Print the agent-facing prompt template for a sponsio workflow.
-
-    Used by the ``sponsio`` skill (``W1``. initial setup, ``W2``.
-    audit & refine) to drive the host agent through contract authoring
-    without burning a separate LLM API call.
-
-    Pair with the corresponding ``--emit-*`` flag:
-
-    \b
-        sponsio onboard . --emit-context     # structured input for prompt
-        sponsio prompt onboard               # the prompt itself
-
-    \b
-        sponsio scan src/ --emit-context
-        sponsio prompt scan
-
-    The agent reads both, applies the prompt to the JSON in its own
-    context, and writes the result via Edit/Write.  No
-    ``UnifiedExtractor`` / API key needed for this path.
-    """
-    from importlib.resources import files
-
-    pkg = files("sponsio.prompts")
-    click.echo(pkg.joinpath(f"{flow}.md").read_text(encoding="utf-8"))
-
-
-# ---------------------------------------------------------------------------
-# main
-# ---------------------------------------------------------------------------
-
-
-def main():
-    cli()
-
-
-if __name__ == "__main__":
-    main()
